### 2025-07-03 b6c9087
attempting tcav
I have functional code, but the concept might be flawed.
