+++ src/ibbi/models/single_class_detection.py
@@ -1,11 +1,15 @@
-# src/ibbi/models/detection.py
+# src/ibbi/models/single_class_detection.py
 
 """
 Single-class beetle object detection models.
 """
 
 import torch
+import torch.nn as nn
 from ultralytics import RTDETR, YOLO, YOLOE, YOLOWorld
+
+# FIX: Import DetectionModel to assist the type checker
+from ultralytics.nn.tasks import DetectionModel
 
 from ..utils.hub import download_from_hf_hub
 from ._registry import register_model
@@ -16,6 +20,9 @@ from ._registry import register_model
 class YOLOSingleClassBeetleDetector:
     """A wrapper class for single-class YOLO beetle detection models."""
 
+    model: YOLO
+    device: str
+
     def __init__(self, model_path: str):
         self.model = YOLO(model_path)
         self.device = "cuda" if torch.cuda.is_available() else "cpu"
@@ -23,12 +30,29 @@ class YOLOSingleClassBeetleDetector:
         self.classes = list(self.model.names.values())
         print(f"YOLO Model loaded on device: {self.device}")
 
+    def to(self, device):
+        """Moves the internal model to the specified device."""
+        self.device = device
+        self.model.to(device)
+        return self
+
     def predict(self, image, **kwargs):
         return self.model.predict(image, **kwargs)
 
     def extract_features(self, image, **kwargs):
-        features = self.model.embed(image, **kwargs)
-        return features[0] if features else None
+        """Extracts feature embeddings. Returns a list of tensors for a batch of images."""
+        kwargs["stream"] = False
+        return self.model.embed(image, **kwargs)
+
+    def predict_from_features(self, features, **kwargs):
+        """Gets classification logits from feature embeddings. Required for TCAV."""
+        # FIX: Assert the type of the internal model to guide pyright
+        internal_model_module = self.model.model
+        assert isinstance(internal_model_module, DetectionModel)
+        if hasattr(internal_model_module, "model") and isinstance(internal_model_module.model, nn.Sequential):
+            detection_head = internal_model_module.model[-1]
+            return detection_head(features)
+        raise NotImplementedError("This YOLO model architecture does not have a standard indexable final layer.")
 
     def get_classes(self) -> list[str]:
         return self.classes
@@ -37,6 +61,9 @@ class YOLOSingleClassBeetleDetector:
 class RTDETRSingleClassBeetleDetector:
     """A wrapper class for single-class RT-DETR beetle detection models."""
 
+    model: RTDETR
+    device: str
+
     def __init__(self, model_path: str):
         self.model = RTDETR(model_path)
         self.device = "cuda" if torch.cuda.is_available() else "cpu"
@@ -44,12 +71,29 @@ class RTDETRSingleClassBeetleDetector:
         self.classes = list(self.model.names.values())
         print(f"RT-DETR Model loaded on device: {self.device}")
 
+    def to(self, device):
+        """Moves the internal model to the specified device."""
+        self.device = device
+        self.model.to(device)
+        return self
+
     def predict(self, image, **kwargs):
         return self.model.predict(image, **kwargs)
 
     def extract_features(self, image, **kwargs):
-        features = self.model.embed(image, **kwargs)
-        return features[0] if features else None
+        """Extracts feature embeddings. Returns a list of tensors for a batch of images."""
+        kwargs["stream"] = False
+        return self.model.embed(image, **kwargs)
+
+    def predict_from_features(self, features, **kwargs):
+        """Gets classification logits from feature embeddings. Required for TCAV."""
+        internal_model_module = self.model.model
+        assert isinstance(internal_model_module, DetectionModel)
+        if hasattr(internal_model_module, "head"):
+            # Add this assertion to confirm the head is a function
[…]
