+++ src/ibbi/xai/tcav.py
@@ -1,0 +1,365 @@
+# src/ibbi/xai/tcav.py
+"""
+ROI-aware TCAV utilities for IBBI multi-class YOLO detectors
+-----------------------------------------------------------
+This re-implementation keeps **spatial information inside each detected
+bounding box** when discovering concepts and computing TCAV scores. The
+pipeline is now:
+
+1. Detect objects -> store **box coordinates** (no cropping).
+2. Forward once per image, capture full neck feature maps.
+3. ROI-align inside each box, global-average-pool per level -> 1 vector/ROI.
+4. Cluster target-class vectors (HDBSCAN) -> concepts.
+5. Learn one CAV per concept (SGD vs. background vectors).
+6. TCAV test: directional derivative of the *box-specific* logit along the CAV.
+
+As a result, every concept and every TCAV score is conditioned on the exact
+pixels of the beetle inside its bounding box, enabling body-part analysis.
+"""
+
+from __future__ import annotations
+
+import math
+import warnings
+from typing import Any
+
+import hdbscan
+import numpy as np
+import torch
+import torch.nn.functional as F
+from PIL import Image
+from sklearn.linear_model import SGDClassifier
+from sklearn.preprocessing import StandardScaler
+from torchvision import transforms
+from torchvision.ops import roi_align
+from tqdm.auto import tqdm
+
+from ibbi.models.multi_class_detection import (
+    YOLOBeetleMultiClassDetector,
+)
+
+warnings.filterwarnings("ignore", category=UserWarning)
+
+# --------------------------------------------------------------
+#           Constants & ROI-pool helper
+# --------------------------------------------------------------
+# Neck strides for YOLOv8/9/10 P3, P4, P5 feature maps.
+STRIDES: list[int] = [8, 16, 32]
+
+
+def _to_tensor_and_pad(pil_img: Image.Image, divisor: int = 32) -> torch.Tensor:
+    """Convert PIL -> Tensor and right/bottom-pad so height & width are multiples of *divisor*."""
+    t = transforms.ToTensor()(pil_img)
+    _, h, w = t.shape
+    new_h = math.ceil(h / divisor) * divisor
+    new_w = math.ceil(w / divisor) * divisor
+    pad_h = new_h - h
+    pad_w = new_w - w
+    if pad_h or pad_w:
+        # pad format: (left, right, top, bottom)
+        t = F.pad(t, (0, pad_w, 0, pad_h))
+    return t
+
+
+def roi_pool_and_flatten(
+    fmap_list: list[torch.Tensor],
+    boxes_xyxy: torch.Tensor,
+) -> torch.Tensor:
+    """ROI-align each feature-map level inside *boxes* and GAP-pool.
+
+    * `fmap_list` - list[C x H x W] feature maps from the neck.
+    * `boxes_xyxy` - `[N,4]` boxes in image-pixel coordinates.
+
+    Returns a tensor `[N, ΣC]` (concatenated pooled vectors).
+    """
+    device = fmap_list[0].device
+    pooled: list[torch.Tensor] = []
+    for lvl, fmap in enumerate(fmap_list):
+        stride = STRIDES[lvl]
+        rois = torch.cat(
+            [
+                torch.zeros(len(boxes_xyxy), 1, device=device),  # batch idx 0
+                boxes_xyxy / stride,
+            ],
+            dim=1,
+        )
+        # 1x1 output gives global avg inside ROI
+        feat = roi_align(
+            fmap.unsqueeze(0),
+            rois,
+            output_size=(1, 1),
+            spatial_scale=1.0,
+            aligned=True,
+        )  # type: ignore
+        pooled.append(feat.flatten(start_dim=2))  # [N, C, 1x1] -> [N, C]
[…]
