### 2025-07-25 5536218
LIME addition with plotting
