+++ src/ibbi/xai/lime.py
@@ -1,0 +1,239 @@
+# src/ibbi/xai/lime.py
+
+"""
+Highly optimized LIME-based model explainability for IBBI models,
+featuring batched predictions and faster segmentation.
+"""
+
+from typing import Callable, Optional, Tuple
+
+import matplotlib.colors as mcolors
+import matplotlib.pyplot as plt
+import numpy as np
+import torch
+from lime import lime_image
+from PIL import Image
+from skimage.segmentation import slic
+from skimage.transform import resize
+
+from ..models import ModelType
+from ..models.zero_shot_detection import GroundingDINOModel
+
+
+def _prediction_wrapper(model: ModelType, text_prompt: Optional[str] = None) -> Callable:
+    """
+    Creates a highly efficient, batched prediction function for LIME.
+    This version processes the entire batch of images as a single tensor.
+    """
+
+    def predict(image_array: np.ndarray) -> np.ndarray:
+        """
+        Batched prediction function for LIME.
+
+        Args:
+            image_array (np.ndarray): A numpy array of perturbed images from LIME,
+                                      with shape (num_images, height, width, channels).
+
+        Returns:
+            np.ndarray: A numpy array of prediction probabilities for each class.
+        """
+        # --- Tensor-based Batching for Maximum Efficiency ---
+        # The model expects a PyTorch tensor in the format (batch, channels, height, width)
+        # and values scaled to [0, 1].
+        # This check is needed because LIME can sometimes send a single image (ndim=3)
+        if image_array.ndim == 3:
+            image_array = np.expand_dims(image_array, 0)
+
+        # 1. Convert NumPy array to a PyTorch tensor.
+        # 2. Permute dimensions from (B, H, W, C) to (B, C, H, W).
+        # 3. Convert to float and scale pixel values from [0, 255] to [0, 1].
+        image_tensor = torch.from_numpy(image_array).permute(0, 3, 1, 2).float() / 255.0
+
+        # Move tensor to the same device as the model for GPU acceleration
+        device = next(model.model.parameters()).device
+        image_tensor = image_tensor.to(device)
+
+        # For GroundingDINOModel (requires a different prediction logic)
+        if isinstance(model, GroundingDINOModel):
+            if not text_prompt:
+                raise ValueError("A 'text_prompt' is required for explaining a GroundingDINOModel.")
+            # This model's predict is not optimized for tensors, so we still use a list
+            images_to_predict = [Image.fromarray(img) for img in image_array]
+            results = model.predict(images_to_predict, text_prompt=text_prompt)
+            num_classes = 1
+            predictions = np.zeros((image_array.shape[0], num_classes))
+            for i, res in enumerate(results):
+                if res["scores"].nelement() > 0:
+                    predictions[i, 0] = res["scores"].max().item()
+
+        # For standard detection models (like YOLO)
+        else:
+            class_names = model.get_classes()
+            num_classes = len(class_names)
+            predictions = np.zeros((image_array.shape[0], num_classes))
+
+            # --- Perform a single, efficient batched prediction on the tensor ---
+            results = model.model(image_tensor)
+
+            # Process the batch of results
+            for i, res in enumerate(results):
+                if hasattr(res, "boxes") and res.boxes is not None:
+                    for box in res.boxes:
+                        class_idx = int(box.cls)
+                        confidence = box.conf.item()
+                        predictions[i, class_idx] = max(predictions[i, class_idx], confidence)
+
+        return predictions
+
+    return predict
+
+
+def explain_with_lime(
+    model: ModelType,
+    image: Image.Image,
+    text_prompt: Optional[str] = None,
[…]
