### 2025-07-25 d481915
Updates bibliography file (#27)
Replaces and updates the bibliography file with corrected and more comprehensive citation information.

Fixes a broken link in the README file.
