+++ paper.bib
@@ -1,62 +1,68 @@
 @book{kirkendall2015,
-  title={Bark Beetles: Biology and Ecology of Native and Invasive Species},
-  author={Kirkendall, L. R. and Biedermann, P. H. and Jordal, B. H.},
-  year={2015},
-  publisher={Academic Press}
+  title     = {Bark Beetles: Biology and Ecology of Native and Invasive Species},
+  author    = {Kirkendall, L. R. and Biedermann, P. H. and Jordal, B. H.},
+  year      = {2015},
+  publisher = {Academic Press},
+  doi       = {10.1016/C2012-0-06455-8}
 }
 
-@article{hulcr2015morphology,
-  title={The morbid anatomy of a pandemic: bark and ambrosia beetle diversity in the age of invasive species},
-  author={Hulcr, Jiri and Atkinson, Thomas H and Cognato, Anthony I and Jordal, Bjarte H and McKenna, Duane D},
-  journal={In: Vega F.E., Hofstetter R.W. (eds) Bark Beetles. Academic Press, London},
-  pages={249--271},
-  year={2015},
-  publisher={Elsevier}
+@incollection{hulcr2015morphology,
+  title     = {The morbid anatomy of a pandemic: bark and ambrosia beetle diversity in the age of invasive species},
+  author    = {Hulcr, Jiri and Atkinson, Thomas H and Cognato, Anthony I and Jordal, Bjarte H and McKenna, Duane D},
+  booktitle = {Bark Beetles: Biology, Ecology, and Management of Native and Invasive Species},
+  editor    = {Vega, Fernando E. and Hofstetter, Richard W.},
+  pages     = {249--271},
+  year      = {2015},
+  publisher = {Academic Press},
+  doi       = {10.1016/B978-0-12-417156-5.00008-9}
 }
 
 @article{valan2019,
-  title={Automated identification of insect species from images},
-  author={Valan, Miroslav and Peri, Eyal and Glinka, Oksana and Zwiebel, Ran and Pupko, Tal and Blank, Lior and Saltz, David},
-  journal={bioRxiv},
-  pages={692336},
-  year={2019},
-  publisher={Cold Spring Harbor Laboratory}
+  title     = {Automated identification of insect species from images},
+  author    = {Valan, Miroslav and Peri, Eyal and Glinka, Oksana and Zwiebel, Ran and Pupko, Tal and Blank, Lior and Saltz, David},
+  journal   = {bioRxiv},
+  year      = {2019},
+  publisher = {Cold Spring Harbor Laboratory},
+  doi       = {10.1101/692336}
 }
 
 @article{hoye2021,
-  title={Deep learning and computer vision will transform entomology},
-  author={Hoye, Toke Thomas and Nater, Alexander and Babil, Pavan K and Velli, Eigil de and filmsbo, Rasmus},
-  journal={Proceedings of the National Academy of Sciences},
-  volume={118},
-  number={2},
-  pages={e2002545117},
-  year={2021},
-  publisher={National Acad Sciences}
+  title     = {Deep learning and computer vision will transform entomology},
+  author    = {Hoye, Toke Thomas and Nater, Alexander and Babil, Pavan K and Velli, Eigil de and filmsbo, Rasmus},
+  journal   = {Proceedings of the National Academy of Sciences},
+  volume    = {118},
+  number    = {2},
+  pages     = {e2023520118},
+  year      = {2021},
+  publisher = {National Academy of Sciences},
+  doi       = {10.1073/pnas.2023520118}
 }
 
-@inproceedings{horn2017,
-  title={The iNaturalist species classification and detection dataset},
[…]
