### 2025-07-25 d3074a7
LIME addition with plotting (#28)
