### 2025-07-25 0e8dedf
uncomment notebooks
