### 2025-07-28 0ee8817
init clustering code
