+++ src/ibbi/cluster/metrics.py
@@ -1,16 +1,18 @@
-import numpy as np
-import pandas as pd
-from sklearn.metrics import (
-    adjusted_rand_score,
-    normalized_mutual_info_score,
-    silhouette_score,
-    davies_bouldin_score,
-    pairwise_distances
-)
-from skbio.stats.distance import mantel
 import importlib.resources
 
-def get_relatedness_matrix(path_to_csv: str = None) -> pd.DataFrame:
+import numpy as np
+import pandas as pd
+from skbio.stats.distance import mantel
+from sklearn.metrics import (
+    adjusted_rand_score,
+    davies_bouldin_score,
+    normalized_mutual_info_score,
+    pairwise_distances,
+    silhouette_score,
+)
+
+
+def get_relatedness_matrix(path_to_csv: str = r"../data/ibbi_species_relatedness_matrix.csv") -> pd.DataFrame:
     """
     Loads the species relatedness matrix from the package data.
 
@@ -39,7 +41,7 @@ def calculate_nmi(labels_true: np.ndarray, labels_pred: np.ndarray) -> float:
     return normalized_mutual_info_score(labels_true, labels_pred)
 
 
-def calculate_silhouette(embeddings: np.ndarray, labels: np.ndarray, metric: str = 'euclidean') -> float:
+def calculate_silhouette(embeddings: np.ndarray, labels: np.ndarray, metric: str = "euclidean") -> float:
     """Calculates the Silhouette Score. Returns -1 if only one cluster is found."""
     if len(np.unique(labels)) < 2:
         return -1.0
@@ -49,15 +51,12 @@ def calculate_silhouette(embeddings: np.ndarray, labels: np.ndarray, metric: str
 def calculate_davies_bouldin(embeddings: np.ndarray, labels: np.ndarray) -> float:
     """Calculates the Davies-Bouldin Index. Returns infinity if only one cluster is found."""
     if len(np.unique(labels)) < 2:
-        return float('inf')
+        return float("inf")
     return davies_bouldin_score(embeddings, labels)
 
 
 def relatedness_correlation_score(
-    embeddings: np.ndarray,
-    species_labels: list,
-    relatedness_matrix: pd.DataFrame,
-    embedding_metric: str = 'cosine'
+    embeddings: np.ndarray, species_labels: list, relatedness_matrix: pd.DataFrame, embedding_metric: str = "cosine"
 ) -> tuple:
     """
     Evaluates clustering by comparing the similarity of embeddings with
@@ -92,11 +91,6 @@ def relatedness_correlation_score(
             genetic_dist_matrix[i, j] = genetic_dist_matrix[j, i] = distance
 
     # 3. Perform the Mantel test
-    corr, p_value, _ = mantel(
-        x=model_dist_matrix,
-        y=genetic_dist_matrix,
-        method='pearson',
-        permutations=999
-    )
+    corr, p_value, _ = mantel(x=model_dist_matrix, y=genetic_dist_matrix, method="pearson", permutations=999)
 
     return corr, p_value
