+++ pyproject.toml
@@ -23,6 +23,8 @@ hf-xet = "^1.1.3"
 shap = "^0.48.0"
 slicer = "^0.0.8"
 lime = "^0.2.0.1"
+hdbscan = "^0.8.40"
+scikit-bio = "^0.7.0"
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.11.10"
