+++ src/ibbi/models/__init__.py
@@ -29,6 +29,12 @@ from .single_class_detection import (
     yolov12x_bb_detect_model,
     yoloworldv2_bb_detect_model,
 )
+from .untrained import (
+    UntrainedFeatureExtractor,
+    convformer_b36_features_model,
+    dinov2_vitl14_lvd142m_features_model,
+    eva02_base_patch14_224_mim_in22k_features_model,
+)
 from .zero_shot_detection import GroundingDINOModel, grounding_dino_detect_model
 
 # --- CORRECTED: Define the ModelType TypeVar ---
@@ -44,6 +50,7 @@ ModelType = TypeVar(
     RTDETRBeetleMultiClassDetector,
     YOLOEBeetleMultiClassDetector,
     GroundingDINOModel,
+    UntrainedFeatureExtractor,
 )
 
 
@@ -68,4 +75,8 @@ __all__ = [
     "yoloe11l_seg_bb_multi_class_detect_model",
     # Zero-Shot Detection Models
     "grounding_dino_detect_model",
+    # untrained feature extractors
+    "dinov2_vitl14_lvd142m_features_model",
+    "eva02_base_patch14_224_mim_in22k_features_model",
+    "convformer_b36_features_model",
 ]
