+++ pyproject.toml
@@ -25,6 +25,7 @@ slicer = "^0.0.8"
 lime = "^0.2.0.1"
 hdbscan = "^0.8.40"
 scikit-bio = "^0.7.0"
+timm = "^1.0.19"
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.11.10"
