### 2025-08-14 #8796 @editorialbot
[REVIEW]: IBBI: A Python package for the detection and classification of bark and ambrosia beetles
**Submitting author:** @ChristopherMarais (<a href="http://orcid.org/0000-0003-3409-0473">Christopher Marais</a>)
**Repository:** https://github.com/ChristopherMarais/IBBI
**Branch with paper.md** (empty if default branch): main
**Version:** v0.2.3
