+++ src/ibbi/evaluate/performance.py
@@ -1,0 +1,219 @@
+from collections import defaultdict
+from typing import Any, Dict, List, Optional, Union
+
+import numpy as np
+from sklearn.metrics import (
+    accuracy_score,
+    classification_report,
+    confusion_matrix,
+    precision_recall_fscore_support,
+)
+
+
+def classification_performance(
+    true_labels: np.ndarray,
+    predicted_labels: np.ndarray,
+    target_names: Optional[List[str]] = None,
+) -> Dict[str, Any]:
+    """
+    Calculates a suite of classification metrics.
+
+    Args:
+        true_labels (np.ndarray): An array of ground-truth labels.
+        predicted_labels (np.ndarray): An array of labels predicted by the model.
+        target_names (List[str], optional): Display names for the labels.
+
+    Returns:
+        Dict[str, Any]: A dictionary containing key performance metrics.
+    """
+    accuracy = accuracy_score(true_labels, predicted_labels)
+    # The type stubs for sklearn are likely incorrect, expecting a str.
+    # "warn" is a valid option that achieves the same goal as 0.
+    precision, recall, f1, _ = precision_recall_fscore_support(
+        true_labels, predicted_labels, average="macro", zero_division="warn"
+    )
+    cm = confusion_matrix(true_labels, predicted_labels)
+    report = classification_report(
+        true_labels,
+        predicted_labels,
+        target_names=target_names,
+        output_dict=True,
+        zero_division="warn",
+    )
+
+    performance_metrics = {
+        "accuracy": accuracy,
+        "macro_precision": precision,
+        "macro_recall": recall,
+        "macro_f1_score": f1,
+        "confusion_matrix": cm.tolist(),
+        "classification_report": report,
+    }
+
+    return performance_metrics
+
+
+def object_detection_performance(
+    gt_boxes: np.ndarray,
+    gt_labels: List[int],
+    gt_image_ids: List[Any],
+    pred_boxes: np.ndarray,
+    pred_labels: List[int],
+    pred_scores: List[float],
+    pred_image_ids: List[Any],
+    iou_thresholds: Union[float, List[float]] = 0.5,
+) -> Dict[str, Any]:
+    """
+    Calculates mean Average Precision (mAP) over one or more IoU thresholds.
+
+    This version accepts flattened lists for detections, where the list indices
+    must correspond (e.g., gt_boxes[i], gt_labels[i], and gt_image_ids[i] all
+    describe the same ground-truth box).
+
+    Args:
+        gt_boxes (np.ndarray): NumPy array of ground-truth bounding boxes with shape (N, 4).
+                                Each row is [x1, y1, x2, y2].
+        gt_labels (List[Any]): List of ground-truth labels [Any,...].
+        gt_image_ids (List[Any]): List of image IDs for each ground-truth box.
+        pred_boxes (np.ndarray): NumPy array of predicted bounding boxes with shape (M, 4).
+                                 Each row is [x1, y1, x2, y2].
+        pred_labels (List[Any]): List of predicted labels [Any,...].
+        pred_scores (List[float]): List of prediction confidence scores [float,...].
+        pred_image_ids (List[Any]): List of image IDs for each predicted box.
+        iou_thresholds (Union[float, List[float]]): A single IoU threshold or a list of thresholds.
+
+    Returns:
+        Dict[str, Any]: A dictionary containing the final averaged mAP, per-class APs for the
+                        last IoU threshold, and mAP scores for each individual threshold.
+    """
+
+    def _calculate_iou(boxA, boxB):
+        # Determine the (x, y)-coordinates of the intersection rectangle
+        xA = max(boxA[0], boxB[0])
[…]
