+++ src/ibbi/evaluate/performance.py
@@ -1,3 +1,5 @@
+# src/ibbi/evaluate/performance.py
+
 from collections import defaultdict
 from typing import Any, Dict, List, Optional, Union
 
@@ -10,6 +12,27 @@ from sklearn.metrics import (
 )
 
 
+def _calculate_iou(boxA, boxB):
+    """Calculates Intersection over Union for two bounding boxes [x1, y1, x2, y2]."""
+    # Determine the (x, y)-coordinates of the intersection rectangle
+    xA = max(boxA[0], boxB[0])
+    yA = max(boxA[1], boxB[1])
+    xB = min(boxA[2], boxB[2])
+    yB = min(boxA[3], boxB[3])
+
+    # Compute the area of intersection
+    interArea = max(0, xB - xA) * max(0, yB - yA)
+
+    # Compute the area of both the prediction and ground-truth rectangles
+    boxAArea = (boxA[2] - boxA[0]) * (boxA[3] - boxA[1])
+    boxBArea = (boxB[2] - boxB[0]) * (boxB[3] - boxB[1])
+
+    # Compute the intersection over union
+    denominator = float(boxAArea + boxBArea - interArea)
+    iou = interArea / denominator if denominator > 0 else 0
+    return iou
+
+
 def classification_performance(
     true_labels: np.ndarray,
     predicted_labels: np.ndarray,
@@ -17,25 +40,21 @@ def classification_performance(
 ) -> Dict[str, Any]:
     """
     Calculates a suite of classification metrics.
-
-    Args:
-        true_labels (np.ndarray): An array of ground-truth labels.
-        predicted_labels (np.ndarray): An array of labels predicted by the model.
-        target_names (List[str], optional): Display names for the labels.
-
-    Returns:
-        Dict[str, Any]: A dictionary containing key performance metrics.
     """
+    labels_lst = list(set(true_labels) | set(predicted_labels))
+    all_labels = target_names if target_names is not None else sorted(labels_lst)
+
     accuracy = accuracy_score(true_labels, predicted_labels)
-    # The type stubs for sklearn are likely incorrect, expecting a str.
-    # "warn" is a valid option that achieves the same goal as 0.
     precision, recall, f1, _ = precision_recall_fscore_support(
-        true_labels, predicted_labels, average="macro", zero_division="warn"
+        true_labels, predicted_labels, average="macro", zero_division="warn", labels=all_labels
     )
-    cm = confusion_matrix(true_labels, predicted_labels)
+
+    cm = confusion_matrix(true_labels, predicted_labels, labels=all_labels)
+
     report = classification_report(
         true_labels,
         predicted_labels,
+        labels=all_labels,
         target_names=target_names,
         output_dict=True,
         zero_division="warn",
@@ -65,58 +84,18 @@ def object_detection_performance(
 ) -> Dict[str, Any]:
     """
     Calculates mean Average Precision (mAP) over one or more IoU thresholds.
-
-    This version accepts flattened lists for detections, where the list indices
-    must correspond (e.g., gt_boxes[i], gt_labels[i], and gt_image_ids[i] all
-    describe the same ground-truth box).
-
-    Args:
-        gt_boxes (np.ndarray): NumPy array of ground-truth bounding boxes with shape (N, 4).
-                                Each row is [x1, y1, x2, y2].
-        gt_labels (List[Any]): List of ground-truth labels [Any,...].
-        gt_image_ids (List[Any]): List of image IDs for each ground-truth box.
-        pred_boxes (np.ndarray): NumPy array of predicted bounding boxes with shape (M, 4).
-                                 Each row is [x1, y1, x2, y2].
-        pred_labels (List[Any]): List of predicted labels [Any,...].
-        pred_scores (List[float]): List of prediction confidence scores [float,...].
[…]
