+++ src/ibbi/evaluate/features.py
@@ -1,5 +1,4 @@
-# MODIFICATION: Added 'cast' to the import from 'typing'
-from typing import TYPE_CHECKING, Dict, List, Optional, Tuple, cast
+from typing import TYPE_CHECKING, Optional, cast
 
 import numpy as np
 import pandas as pd
@@ -40,6 +39,8 @@ if TYPE_CHECKING:
     import umap
     from skbio.stats.distance import mantel
 
+from importlib import resources as pkg_resources
+
 
 class EmbeddingEvaluator:
     """
@@ -51,9 +52,6 @@ class EmbeddingEvaluator:
     evaluation metrics.
     """
 
-    # MODIFICATION: Added explicit type annotations for instance attributes.
-    # This solves the error on line 212 by ensuring the type checker
-    # always knows `processed_data` is a NumPy array that can be indexed.
     embeddings: np.ndarray
     processed_data: np.ndarray
     predicted_labels: np.ndarray
@@ -162,7 +160,7 @@ class EmbeddingEvaluator:
         self.predicted_labels = clusterer.fit_predict(self.processed_data)
         print("HDBSCAN clustering complete.")
 
-    def evaluate_against_truth(self, true_labels: np.ndarray) -> Dict[str, float]:
+    def evaluate_against_truth(self, true_labels: np.ndarray) -> dict[str, float]:
         """
         Calculates external clustering validation metrics against ground truth labels.
 
@@ -176,7 +174,7 @@ class EmbeddingEvaluator:
                 array must be the same length as the original embeddings.
 
         Returns:
-            Dict[str, float]: A dictionary containing the calculated scores:
+            dict[str, float]: A dictionary containing the calculated scores:
                 - "ARI": The Adjusted Rand Index, ranging from -1 (bad) to 1 (perfect).
                 - "NMI": The Normalized Mutual Information, ranging from 0 (bad) to 1 (perfect).
         """
@@ -184,7 +182,7 @@ class EmbeddingEvaluator:
         nmi = normalized_mutual_info_score(true_labels, self.predicted_labels)
         return {"ARI": ari, "NMI": nmi}
 
-    def evaluate_cluster_structure(self) -> Dict[str, float]:
+    def evaluate_cluster_structure(self) -> dict[str, float]:
         """
         Calculates internal clustering validation metrics based on cluster structure.
 
@@ -194,7 +192,7 @@ class EmbeddingEvaluator:
         excluded from these calculations.
 
         Returns:
-            Dict[str, float]: A dictionary containing the calculated scores:
+            dict[str, float]: A dictionary containing the calculated scores:
                 - "Silhouette_Score": Measures how similar an object is to its
                   own cluster compared to other clusters. Ranges from -1 to 1,
                   where higher is better.
@@ -220,10 +218,10 @@ class EmbeddingEvaluator:
         self,
         true_labels: np.ndarray,
         embedding_metric: str = "cosine",
-        ext_dist_path: str = "../data/ibbi_species_distance_matrix.csv",
+        # The ext_dist_path parameter is removed as it's no longer used.
         ext_dist_matrix: Optional[np.ndarray] = None,
-        ext_dist_labels: Optional[List[str]] = None,
-    ) -> Tuple[float, float, int]:
+        ext_dist_labels: Optional[list[str]] = None,
+    ) -> tuple[float, float, int]:
         """
         Calculates Mantel correlation between embedding distances and an external distance matrix.
 
@@ -234,24 +232,22 @@ class EmbeddingEvaluator:
         The external distance matrix can be provided in two ways:
         1. As a CSV file via `ext_dist_path` (default behavior).
         2. Directly as a NumPy array and a list of labels via `ext_dist_matrix` and
-           `ext_dist_labels`.
+        `ext_dist_labels`.
 
         Args:
[…]
