### 2025-08-21 64a5cda
Feature/global xai clustering (#30)
* uncomment notebooks

* added relatedness matrix

* init clustering code v1.0

* pychace ignored

* Auto stash before merge of "feature/global_xai_clustering" and "origin/feature/global_xai_clustering"

* performance and feature clustering

* states of models fix & pytorch fix
