+++ environment.yml
@@ -1,9 +1,10 @@
 name: IBBI
 channels:
+  - nvidia
   - conda-forge
   - defaults
 dependencies:
   - python=3.11
   - pip
   - poetry
-  - cudatoolkit
+  - cuda-toolkit
