### 2025-08-22 167ecc3
replacing hdbscan with sklearn.cluster.HDBSCAN (#31)
