### 2025-08-26 #8796 @goldingn
Hi @martibosch @PetervanLunteren, please create your individual review checklists by typing:
`@editorialbot generate my checklist`

Then you can work through those items checking them off as you conduct your review over the coming weeks
