+++ src/ibbi/models/multi_class_detection.py
@@ -5,12 +5,10 @@ Multi-class beetle object detection models.
 """
 
 import torch
-from ultralytics import RTDETR, YOLO, YOLOE
+from ultralytics import RTDETR, YOLO
 
 from ..utils.hub import download_from_hf_hub
 from ._registry import register_model
-
-# --- Base Classes for different architectures ---
 
 
 class YOLOBeetleMultiClassDetector:
@@ -53,30 +51,6 @@ class RTDETRBeetleMultiClassDetector:
 
     def get_classes(self) -> list[str]:
         return self.classes
-
-
-class YOLOEBeetleMultiClassDetector:
-    """A wrapper class for YOLOE multi-class beetle detector models."""
-
-    def __init__(self, model_path: str):
-        self.model = YOLOE(model_path)
-        self.device = "cuda" if torch.cuda.is_available() else "cpu"
-        self.model.to(self.device)
-        self.classes = list(self.model.names.values())
-        print(f"YOLOE Multi-Class Detector Model loaded on device: {self.device}")
-
-    def predict(self, image, **kwargs):
-        return self.model.predict(image, **kwargs)
-
-    def extract_features(self, image, **kwargs):
-        features = self.model.embed(image, **kwargs)
-        return features[0] if features else None
-
-    def get_classes(self) -> list[str]:
-        return self.classes
-
-
-# --- Factory Functions for Multi-Class Detection Models ---
 
 
 @register_model
@@ -143,9 +117,3 @@ def rtdetrx_bb_multi_class_detect_model(pretrained: bool = False, **kwargs):
     else:
         local_weights_path = "rtdetr-x.pt"
     return RTDETRBeetleMultiClassDetector(model_path=local_weights_path)
-
-
-@register_model
-def yoloe11l_seg_bb_multi_class_detect_model(pretrained: bool = False, **kwargs):
-    local_weights_path = "yoloe-11l-seg.pt"
-    return YOLOEBeetleMultiClassDetector(model_path=local_weights_path)
