### 2025-08-28 e0a3cd8
improved caching
