+++ src/ibbi/utils/data.py
@@ -9,7 +9,7 @@ from pathlib import Path
 from typing import Union
 
 from datasets import Dataset, DatasetDict, IterableDataset, IterableDatasetDict, load_dataset
-from huggingface_hub import hf_hub_download
+from huggingface_hub import hf_hub_download, snapshot_download
 from PIL import Image
 
 # Import the cache utility to manage download locations
@@ -18,19 +18,21 @@ from .cache import get_cache_dir
 
 def get_dataset(
     repo_id: str = "IBBI-bio/ibbi_test_data",
+    local_dir: str = "ibbi_test_data",
     split: str = "train",
     **kwargs,
 ) -> Dataset:
     """
-    Loads a dataset from the Hugging Face Hub.
+    Downloads and loads a dataset from the Hugging Face Hub into a specified local directory.
 
-    This function is a wrapper around `datasets.load_dataset` and returns
-    the raw Dataset object, allowing for direct manipulation. By default,
-    it downloads data to a local '.cache' directory where the script is run.
+    This function ensures the dataset is stored in a folder with a clean name.
+    If the dataset already exists locally, the download is skipped.
 
     Args:
         repo_id (str): The Hugging Face Hub repository ID of the dataset.
                          Defaults to "IBBI-bio/ibbi_test_data".
+        local_dir (str): The desired local directory name to store the dataset.
+                         Defaults to "ibbi_test_data".
         split (str): The dataset split to use (e.g., "train", "test").
                          Defaults to "train".
         **kwargs: Additional keyword arguments passed directly to
@@ -41,29 +43,23 @@ def get_dataset(
 
     Raises:
         TypeError: If the loaded object is not of type `Dataset`.
-
-    Example:
-        ```python
-        import ibbi
-
-        # Load the default test dataset
-        test_data = ibbi.get_dataset()
-
-        # Iterate through the first 5 examples
-        for i, example in enumerate(test_data):
-            if i >= 5:
-                break
-            print(example['image'])
-        ```
     """
-    print(f"Loading dataset '{repo_id}' into local directory...")
+    dataset_path = Path(local_dir)
+
+    if not dataset_path.exists():
+        print(f"Dataset not found locally. Downloading from '{repo_id}' to '{dataset_path}'...")
+        snapshot_download(
+            repo_id=repo_id, repo_type="dataset", local_dir=str(dataset_path), local_dir_use_symlinks=False
+        )
+        print("Download complete.")
+    else:
+        print(f"Found cached dataset at '{dataset_path}'. Loading from disk.")
+
     try:
-        # '.cache' folder relative to the script's execution path.
         dataset: Union[Dataset, DatasetDict, IterableDataset, IterableDatasetDict] = load_dataset(
-            repo_id, split=split, trust_remote_code=True, cache_dir=".", **kwargs
+            str(dataset_path), split=split, trust_remote_code=True, **kwargs
         )
 
-        # Ensure that the returned object is a Dataset
         if not isinstance(dataset, Dataset):
             raise TypeError(
                 f"Expected a 'Dataset' object for split '{split}', but received type '{type(dataset).__name__}'."
@@ -72,50 +68,53 @@ def get_dataset(
         print("Dataset loaded successfully.")
         return dataset
     except Exception as e:
-        print(f"Failed to load dataset '{repo_id}'. Please check the repository ID and your connection.")
+        print(f"Failed to load dataset from '{dataset_path}'. Please check the path and your connection.")
         raise e
 
 
-def get_shap_background_dataset() -> list[dict]:
+def get_shap_background_dataset(image_size: tuple[int, int] = (224, 224)) -> list[dict]:
     """
[…]
