### 2025-08-29 faa3fec
added DINOv3 feature extractor
