+++ src/ibbi/models/untrained.py
@@ -4,11 +4,13 @@
 Untrained models for feature extraction.
 """
 
+import numpy as np
 import timm
 import torch
 from PIL import Image
 from timm.data.config import resolve_model_data_config
 from timm.data.transforms_factory import create_transform
+from transformers import pipeline
 
 from ._registry import register_model
 
@@ -35,16 +37,51 @@ class UntrainedFeatureExtractor:
         else:
             raise TypeError("Image must be a PIL Image or a file path.")
 
-        # Prepare the input tensor. The create_transform function has a complex
-        # return type (Union) that pyright struggles to resolve, so we ignore the type error here.
-        input_tensor = self.transforms(img).unsqueeze(0).to(self.device)  # type: ignore
+        # Add a check to ensure self.transforms is a single callable function
+        if not callable(self.transforms):
+            # This error will be raised if it's unexpectedly a tuple
+            raise TypeError("The transform object is not callable. Check the 'separate' argument in create_transform.")
 
-        # Pyright cannot statically determine that the dynamically created timm model has
-        # forward_features and forward_head methods, so we ignore the type errors.
+        # Now, Pylance knows self.transforms is callable and the error will disappear.
+        transformed_img = self.transforms(img)
+
+        # Explicitly ensure the output is a tensor before further operations
+        input_tensor = torch.as_tensor(transformed_img).unsqueeze(0).to(self.device)
+
         features = self.model.forward_features(input_tensor)  # type: ignore
         output = self.model.forward_head(features, pre_logits=True)  # type: ignore
 
         return output.detach().cpu().numpy()
+
+    def get_classes(self) -> list[str]:
+        raise NotImplementedError("This model is for feature extraction only and does not have classes.")
+
+
+class HuggingFaceFeatureExtractor:
+    """A wrapper class for using pretrained Hugging Face models for feature extraction."""
+
+    def __init__(self, model_name: str):
+        self.feature_extractor = pipeline(task="image-feature-extraction", model=model_name)
+        print(f"{model_name} model loaded successfully using the pipeline.")
+
+    def predict(self, image, **kwargs):
+        raise NotImplementedError("This model is for feature extraction only and does not support prediction.")
+
+    def extract_features(self, image, **kwargs):
+        if isinstance(image, str):
+            img = Image.open(image).convert("RGB")
+        elif isinstance(image, Image.Image):
+            img = image
+        else:
+            raise TypeError("Image must be a PIL Image or a file path.")
+
+        # The pipeline returns a list containing a numpy array of all token embeddings.
+        embedding = self.feature_extractor(img, **kwargs)
+
+        # This gives us a single, representative feature vector for the image.
+        global_features = np.array(embedding)[0, 0, :]
+
+        return global_features
 
     def get_classes(self) -> list[str]:
         raise NotImplementedError("This model is for feature extraction only and does not have classes.")
@@ -63,3 +100,9 @@ def eva02_base_patch14_224_mim_in22k_features_model(pretrained: bool = True, **k
 @register_model
 def convformer_b36_features_model(pretrained: bool = True, **kwargs):
     return UntrainedFeatureExtractor(model_name="caformer_b36.sail_in22k_ft_in1k_384")
+
+
+@register_model
+def dinov3_vitl16_lvd1689m_features_model(pretrained: bool = True, **kwargs):
+    # Use the HuggingFaceFeatureExtractor with the IBBI-bio model
+    return HuggingFaceFeatureExtractor(model_name="IBBI-bio/dinov3-vitl16-pretrain-lvd1689m")
