+++ pyproject.toml
@@ -19,14 +19,14 @@ priority = "explicit"
 
 [tool.poetry.dependencies]
 python = ">=3.11,<4.0"
-huggingface-hub = ">=0.32.3,<0.33.0"
+huggingface-hub = "^0.34.4"
 ultralytics = "8.3.139"
 pandas = "^2.3.0"
 numpy = "^2.2.6"
 pillow = "^11.2.1"
 ipywidgets = "^8.1.7"
 datasets = "^3.6.0"
-transformers = "^4.52.4"
+transformers = "^4.56.0"
 hf-xet = "^1.1.3"
 shap = "^0.48.0"
 slicer = "^0.0.8"
