### 2025-09-01 474fada
Feature/xai shap dataset (#32)
* shap background dataset added

* improved caching

* shap soft masking added

* added DINOv3 feature extractor

* update model summary table
