### 2025-09-01 5b6e8b4
update zero shot models to include the same functionality as other models (#33)
