### 2025-09-02 c6700cb
yoloworld extract pred, confidence fix
