+++ pyproject.toml
@@ -37,6 +37,7 @@ umap-learn = "^0.5.9.post2"
 hdbscan = "^0.8.40"
 torch = { version = ">=2.1", source = "pytorch_cu128" }
 torchvision = { version = ">=0.16.0", source = "pytorch_cu128" }
+clip = {git = "https://github.com/openai/CLIP.git"}
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.11.10"
