### 2025-09-04 ChristopherMarais/IBBI#34 @PetervanLunteren
JOSS review revisions
Hi @ChristopherMarais,

I'm Peter one of the reviewers for the [JOSS reveiw](https://github.com/openjournals/joss-reviews/issues/8796#issuecomment-3240165937). I've had a look at the project and great work! Impressive.

There are a few things that I would like to note before I can finish [my review](https://github.com/openjournals/joss-reviews/issues/8796#issuecomment-3254096693).

**Reproducibility**
Even though the manuscript mentions benchmarks, I could not find any quantitative performance metrics. It would be great if we could have a table with metrics per model for accuracy/precision/recall/F1 scores.

**Functionality**
I haven't had the time to test that properly. Will get back on that soon.

**Automated Tests**
I could not find an automated test. This can be as simple as run this CLI command on this image and if it produces these results, the installation is working as expected!

**State of the field**
The paper claims to be novel ("first software package to...") but doesn't cite existing competing tools. Are there none at all? Or are there some that are slightly in the same area? There are a number of open-source insect models available, right? How do these differ from IBBI?

**Quality of writing**
There were a number of typo's found.
Line 84: "suopports" → "supports"
Line 91: "fro" → "for"
Line 93: Code block comment is cut off
Line 104: Incomplete sentence: "torch.Tensor feature embedding of t"
Line 110: Inconsistent capitalization
Line 132: Placeholder "IMPORTANT TO MENTION FUNDING BODIES" must be completed

**References**
Line 42: Incomplete iNaturalist citation "(?)"

**Other notes**
- You mention "object detection and classification models" (Line 66). Can you elaborate which classification models. In the Core tasks and API part, you dont mention any classification models.
