### 2025-09-05 ChristopherMarais/IBBI#35 @PetervanLunteren
Quick start fails on zero-shot detection on Apple Silicon M1
When running your [quick start python code](https://github.com/ChristopherMarais/IBBI/tree/main?tab=readme-ov-file#quick-start), the install fails on the zero-shot detection. The multi-class species detection, and single-class detection work as expected. See output below. It seems like the `zs_detector.predict(image_source, text_prompt="insect")` does not return any predictions (see added debug print statement).

```python

DEBUG - zs_results: {'scores': tensor([]), 'boxes': tensor([], size=(0, 4)), 'text_labels': [], 'labels': []}

--- Zero-Shot Detection (prompt: 'insect') ---
Traceback (most recent call last):
  File "/Users/peter/Desktop/JOSS-review.py", line 41, in <module>
    for box in zs_results[0].boxes:
               ~~~~~~~~~~^^^
  File "/Users/peter/miniforge3/envs/iibi-env/lib/python3.11/site-packages/transformers/models/grounding_dino/processing_grounding_dino.py", line 95, in __getitem__
    return super().__getitem__(key)
           ^^^^^^^^^^^^^^^^^^^^^^^^
KeyError: 0
```

I'm on an Apple Silicon M1 Macbook. I noticed that there was a warning when installing `ibbi`, stating that it would be best to use PEP 517. I've tried both, and both have the same result.

```bash
pip install ibbi
pip install ibbi --use-pep517
```

For reference purposes, this is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/8796
