### 2025-09-16 17d8e36
package layout: renamed modules and classes for usability
