+++ src/ibbi/__init__.py
@@ -2,91 +2,71 @@
 
 """
 Main initialization file for the ibbi package.
+
+Provides a high-level API for model creation, evaluation, and explanation.
 """
 
 from typing import Any
 
-from .evaluate.features import EmbeddingEvaluator
-from .evaluate.performance import classification_performance, object_detection_performance
-from .explain.lime import explain_with_lime, plot_lime_explanation
-from .explain.shap import explain_with_shap, plot_shap_explanation
+# --- Core Functionality ---
+from .datasets import get_dataset, get_shap_background_dataset
 
-# Import model modules to ensure registry is populated
-# --- Import ModelType from its new location ---
-from .models import (
-    ModelType,
-    multi_class_detection,  # noqa: F401
-    single_class_detection,  # noqa: F401
-    untrained,  # noqa: F401
-    zero_shot_detection,  # noqa: F401
-)
-
-# Import the populated registry
+# --- High-level classes for streamlined workflow ---
+from .evaluate import Evaluator
+from .explain import Explainer
+from .models import ModelType
 from .models._registry import model_registry
 from .utils.cache import clean_cache, get_cache_dir
-from .utils.data import get_dataset, get_shap_background_dataset
-
-# --- Top-level function imports ---
 from .utils.info import list_models
+
+# --- Model Aliases for User Convenience ---
+MODEL_ALIASES = {
+    "beetle_detector": "yolov10x_bb_detect_model",
+    "species_classifier": "yolov12x_bb_multi_class_detect_model",
+    "feature_extractor": "dinov3_vitl16_lvd1689m_features_model",
+    "zero_shot_detector": "grounding_dino_detect_model",
+}
 
 
 def create_model(model_name: str, pretrained: bool = False, **kwargs: Any) -> ModelType:
     """
-    Creates a model from a name.
+    Creates a model from a name or a task-based alias.
 
-    This factory function is the main entry point for users of the package.
-    It looks up the requested model in the registry, downloads pretrained
-    weights from the Hugging Face Hub if requested, and returns an
-    instantiated model object.
+    This is the main entry point for creating models. It can use a specific
+    model name or a simplified alias like "species_classifier".
 
     Args:
-        model_name (str): Name of the model to create.
-        pretrained (bool): Whether to load pretrained weights from the Hugging Face Hub.
-                            Defaults to False.
-        **kwargs (Any): Extra arguments to pass to the model-creating function.
+        model_name (str): Name or alias of the model to create.
+        pretrained (bool): Whether to load pretrained weights. Defaults to False.
+        **kwargs (Any): Extra arguments for the model function.
 
     Returns:
-        ModelType: An instance of the requested model (e.g., YOLOSingleClassBeetleDetector or
-                   YOLOBeetleMultiClassDetector).
-
-    Raises:
-        KeyError: If the requested `model_name` is not found in the model registry.
-
-    Example:
-        ```python
-        import ibbi
-
-        # Create a pretrained single-class detection model
-        detector = ibbi.create_model("yolov10x_bb_detect_model", pretrained=True)
-
-        # Create a pretrained multi-class detection model
-        multi_class_detector = ibbi.create_model("yolov10x_bb_multi_class_detect_model", pretrained=True)
-        ```
+        An instance of the requested model.
     """
+    # Resolve alias if used
+    if model_name in MODEL_ALIASES:
+        model_name = MODEL_ALIASES[model_name]
+
     if model_name not in model_registry:
         available = ", ".join(model_registry.keys())
-        raise KeyError(f"Model '{model_name}' not found. Available models: [{available}]")
+        aliases = ", ".join(MODEL_ALIASES.keys())
[…]
