+++ src/ibbi/models/__init__.py
@@ -2,8 +2,16 @@
 
 from typing import TypeVar
 
-# --- Import the actual model classes ---
-from .multi_class_detection import (
+# --- Import all model classes and factory functions to populate the registry ---
+from .feature_extractors import (
+    HuggingFaceFeatureExtractor,
+    UntrainedFeatureExtractor,
+    convformer_b36_features_model,
+    dinov2_vitl14_lvd142m_features_model,
+    dinov3_vitl16_lvd1689m_features_model,
+    eva02_base_patch14_224_mim_in22k_features_model,
+)
+from .multi_class import (
     RTDETRBeetleMultiClassDetector,
     YOLOBeetleMultiClassDetector,
     rtdetrx_bb_multi_class_detect_model,
@@ -13,7 +21,7 @@ from .multi_class_detection import (
     yolov11x_bb_multi_class_detect_model,
     yolov12x_bb_multi_class_detect_model,
 )
-from .single_class_detection import (
+from .single_class import (
     RTDETRSingleClassBeetleDetector,
     YOLOSingleClassBeetleDetector,
     rtdetrx_bb_detect_model,
@@ -23,20 +31,14 @@ from .single_class_detection import (
     yolov11x_bb_detect_model,
     yolov12x_bb_detect_model,
 )
-from .untrained import (
-    UntrainedFeatureExtractor,
-    convformer_b36_features_model,
-    dinov2_vitl14_lvd142m_features_model,
-    dinov3_vitl16_lvd1689m_features_model,
-    eva02_base_patch14_224_mim_in22k_features_model,
-)
-from .zero_shot_detection import (
+from .zero_shot import (
     GroundingDINOModel,
     YOLOWorldModel,
     grounding_dino_detect_model,
     yoloworldv2_bb_detect_model,
 )
 
+# --- Define a Generic ModelType for type hinting ---
 ModelType = TypeVar(
     "ModelType",
     YOLOSingleClassBeetleDetector,
@@ -46,11 +48,20 @@ ModelType = TypeVar(
     GroundingDINOModel,
     YOLOWorldModel,
     UntrainedFeatureExtractor,
+    HuggingFaceFeatureExtractor,
 )
 
-
+# --- Explicitly define the public API of this module ---
 __all__ = [
+    "GroundingDINOModel",
+    "HuggingFaceFeatureExtractor",
     "ModelType",
+    "RTDETRBeetleMultiClassDetector",
+    "RTDETRSingleClassBeetleDetector",
+    "UntrainedFeatureExtractor",
+    "YOLOBeetleMultiClassDetector",
+    "YOLOSingleClassBeetleDetector",
+    "YOLOWorldModel",
     "convformer_b36_features_model",
     "dinov2_vitl14_lvd142m_features_model",
     "dinov3_vitl16_lvd1689m_features_model",

[…]
