### 2025-09-17 52b3dc6
added new example notebooks for new layout
