+++ src/ibbi/models/multi_class.py
@@ -1,4 +1,4 @@
-# src/ibbi/models/multi_class_detection.py
+# src/ibbi/models/multi_class.py
 
 """
 Multi-class beetle object detection models.
@@ -22,7 +22,17 @@ class YOLOBeetleMultiClassDetector:
         print(f"YOLO Multi-Class Detector Model loaded on device: {self.device}")
 
     def predict(self, image, **kwargs):
-        return self.model.predict(image, **kwargs)
+        results = self.model.predict(image, **kwargs)
+
+        result_dict = {"scores": [], "labels": [], "boxes": []}
+
+        if results and hasattr(results[0], "boxes") and results[0].boxes is not None:
+            for box in results[0].boxes:
+                result_dict["scores"].append(box.conf.item())
+                result_dict["labels"].append(self.model.names[int(box.cls)])
+                result_dict["boxes"].append(box.xyxy[0].tolist())
+
+        return result_dict
 
     def extract_features(self, image, **kwargs):
         features = self.model.embed(image, **kwargs)
@@ -43,7 +53,17 @@ class RTDETRBeetleMultiClassDetector:
         print(f"RT-DETR Multi-Class Detector Model loaded on device: {self.device}")
 
     def predict(self, image, **kwargs):
-        return self.model.predict(image, **kwargs)
+        results = self.model.predict(image, **kwargs)
+
+        result_dict = {"scores": [], "labels": [], "boxes": []}
+
+        if results and hasattr(results[0], "boxes") and results[0].boxes is not None:
+            for box in results[0].boxes:
+                result_dict["scores"].append(box.conf.item())
+                result_dict["labels"].append(self.model.names[int(box.cls)])
+                result_dict["boxes"].append(box.xyxy[0].tolist())
+
+        return result_dict
 
     def extract_features(self, image, **kwargs):
         features = self.model.embed(image, **kwargs)
