+++ src/ibbi/evaluate/__init__.py
@@ -7,10 +7,11 @@ Provides the high-level Evaluator class for assessing model performance.
 from typing import Any, Optional, Union
 
 import numpy as np
-import torch
 from tqdm import tqdm
 
 from ..models import ModelType
+from ..models.feature_extractors import HuggingFaceFeatureExtractor
+from ..models.zero_shot import GroundingDINOModel
 from .classification import classification_performance
 from .embeddings import EmbeddingEvaluator
 from .object_detection import object_detection_performance
@@ -32,27 +33,59 @@ class Evaluator:
             predict_kwargs = {}
 
         print("Running classification evaluation...")
-        true_labels = [item["label"] for item in dataset]
+
+        if isinstance(self.model, (HuggingFaceFeatureExtractor, GroundingDINOModel)):
+            print("Warning: Classification evaluation is not supported for this model type.")
+            return {}
+
+        if not hasattr(self.model, "get_classes") or not callable(self.model.get_classes):
+            print("Warning: Model does not have a 'get_classes' method for class mapping. Skipping classification.")
+            return {}
+
+        model_classes = self.model.get_classes()
+        class_name_to_idx = {v: k for k, v in enumerate(model_classes)}
+        true_labels = []
+        items_for_prediction = []
+
+        for item in dataset:
+            if "objects" in item and "category" in item["objects"] and item["objects"]["category"]:
+                label_name = item["objects"]["category"][0]
+                if label_name in class_name_to_idx:
+                    true_labels.append(class_name_to_idx[label_name])
+                    items_for_prediction.append(item)
+
+        if not true_labels:
+            print(
+                "Warning: No valid labels found in the dataset that match the model's classes. Skipping classification."
+            )
+            return {}
+
         predicted_labels = []
 
         print("Making predictions for classification report...")
-        for item in tqdm(dataset):
+        for i, item in enumerate(tqdm(items_for_prediction)):
             results = self.model.predict(item["image"], verbose=False, **predict_kwargs)
-            if not results:
+            true_label_for_item = true_labels[i]
+
+            if not results or not results.get("labels"):
                 predicted_labels.append(-1)
                 continue
 
-            # Check if boxes exist before trying to access them
-            pred_classes = {int(box.cls) for res in results if res.boxes is not None for box in res.boxes}
+            pred_labels_for_item = results.get("labels", [])
+            pred_classes = {class_name_to_idx[label] for label in pred_labels_for_item if label in class_name_to_idx}
 
-            if item["label"] in pred_classes:
-                predicted_labels.append(item["label"])
-            elif pred_classes:
-                # Fallback to the highest confidence prediction
-                all_boxes = [box for res in results if res.boxes is not None for box in res.boxes]
-                if all_boxes:
-                    highest_conf_box = max(all_boxes, key=lambda b: b.conf.item())
-                    predicted_labels.append(int(highest_conf_box.cls))
+            if true_label_for_item in pred_classes:
+                predicted_labels.append(true_label_for_item)
+            elif pred_labels_for_item:
+                # Find the label with the highest score
+                scores = results.get("scores", [])
+                if scores:
+                    highest_conf_idx = np.argmax(scores)
+                    highest_conf_label = pred_labels_for_item[highest_conf_idx]
+                    if highest_conf_label in class_name_to_idx:
+                        predicted_labels.append(class_name_to_idx[highest_conf_label])
+                    else:
+                        predicted_labels.append(-1)
                 else:
                     predicted_labels.append(-1)
             else:
@@ -70,33 +103,52 @@ class Evaluator:
             predict_kwargs = {}
 
[…]
