+++ src/ibbi/evaluate/__init__.py
@@ -1,9 +1,5 @@
 # src/ibbi/evaluate/__init__.py
 
-"""
-Provides the high-level Evaluator class for assessing model performance.
-"""
-
 from typing import Any, Optional, Union
 
 import numpy as np
@@ -42,7 +38,11 @@ class Evaluator:
             print("Warning: Model does not have a 'get_classes' method for class mapping. Skipping classification.")
             return {}
 
-        model_classes = self.model.get_classes()
+        raw_model_classes = self.model.get_classes()
+        if isinstance(raw_model_classes, dict):
+            model_classes = list(raw_model_classes.values())
+        else:
+            model_classes = raw_model_classes
         class_name_to_idx = {v: k for k, v in enumerate(model_classes)}
         true_labels = []
         items_for_prediction = []
@@ -91,7 +91,11 @@ class Evaluator:
             else:
                 predicted_labels.append(-1)
 
-        return classification_performance(np.array(true_labels), np.array(predicted_labels), **kwargs)
+        # This removes the duplicate 'target_names' argument from kwargs if it exists
+        kwargs.pop("target_names", None)
+        return classification_performance(
+            np.array(true_labels), np.array(predicted_labels), target_names=model_classes, **kwargs
+        )
 
     def object_detection(
         self, dataset, iou_thresholds: Union[float, list[float]] = 0.5, predict_kwargs: Optional[dict[str, Any]] = None
@@ -112,8 +116,13 @@ class Evaluator:
             print("Warning: Model does not have a 'get_classes' method for class mapping. Skipping object detection.")
             return {}
 
-        model_classes = self.model.get_classes()
+        raw_model_classes = self.model.get_classes()
+        if isinstance(raw_model_classes, dict):
+            model_classes: list[str] = list(raw_model_classes.values())
+        else:
+            model_classes: list[str] = raw_model_classes
         class_name_to_idx = {v: k for k, v in enumerate(model_classes)}
+        idx_to_name = dict(enumerate(model_classes))
 
         gt_boxes, gt_labels, gt_image_ids = [], [], []
         pred_boxes, pred_labels, pred_scores, pred_image_ids = [], [], [], []
@@ -125,7 +134,6 @@ class Evaluator:
                     label_name = item["objects"]["category"][j]
                     if label_name in class_name_to_idx:
                         bbox = item["objects"]["bbox"][j]
-                        # Convert [x, y, width, height] to [x1, y1, x2, y2]
                         x1, y1, w, h = bbox
                         x2 = x1 + w
                         y2 = y1 + h
@@ -144,7 +152,7 @@ class Evaluator:
                     if label in class_name_to_idx:
                         pred_labels.append(class_name_to_idx[label])
                     else:
-                        pred_labels.append(-1)  # Or handle unknown labels as needed
+                        pred_labels.append(-1)
                     pred_scores.append(score)
                     pred_image_ids.append(i)
 
@@ -159,14 +167,17 @@ class Evaluator:
             iou_thresholds=iou_thresholds,
         )
 
-        # Map integer class IDs in the results back to class names for the final output
         if "per_class_AP_at_last_iou" in performance_results:
             class_aps = performance_results["per_class_AP_at_last_iou"]
-            idx_to_class_name = dict(enumerate(model_classes))
             named_class_aps = {
-                idx_to_class_name.get(class_id, f"unknown_class_{class_id}"): ap for class_id, ap in class_aps.items()
+                idx_to_name.get(class_id, f"unknown_class_{class_id}"): ap for class_id, ap in class_aps.items()
             }
             performance_results["per_class_AP_at_last_iou"] = named_class_aps
+
+        if "sample_results" in performance_results:
+            performance_results["sample_results"]["label"] = performance_results["sample_results"]["label"].map(
+                idx_to_name
+            )
 
[…]
