+++ src/ibbi/evaluate/embeddings.py
@@ -9,10 +9,13 @@ from scipy.spatial.distance import pdist, squareform
 from sklearn.cluster import HDBSCAN
 from sklearn.metrics import (
     adjusted_rand_score,
+    calinski_harabasz_score,
     davies_bouldin_score,
     normalized_mutual_info_score,
     silhouette_score,
+    v_measure_score,
 )
+from sklearn.preprocessing import LabelEncoder
 
 # --- Library Import Checks ---
 try:
@@ -32,6 +35,14 @@ except ImportError:
 if TYPE_CHECKING:
     import umap
     from skbio.stats.distance import mantel
+
+
+def _cluster_purity(y_true, y_pred):
+    """
+    Calculates cluster purity.
+    """
+    contingency_matrix = np.histogram2d(y_true, y_pred, bins=(len(np.unique(y_true)), len(np.unique(y_pred))))[0]
+    return np.sum(np.amax(contingency_matrix, axis=0)) / np.sum(contingency_matrix)
 
 
 class EmbeddingEvaluator:
@@ -57,6 +68,7 @@ class EmbeddingEvaluator:
         min_samples: Optional[int] = None,
         cluster_selection_epsilon: float = 0.0,
         hdbscan_metric: str = "euclidean",
+        allow_single_cluster: bool = False,
         random_state: int = 42,
     ):
         """
@@ -92,41 +104,89 @@ class EmbeddingEvaluator:
             min_samples=min_samples,
             cluster_selection_epsilon=cluster_selection_epsilon,
             metric=clustering_metric,
+            allow_single_cluster=allow_single_cluster,
         )
         self.predicted_labels = clusterer.fit_predict(self.processed_data)
         print("HDBSCAN clustering complete.")
 
-    def evaluate_against_truth(self, true_labels: np.ndarray) -> dict[str, float]:
+    def get_sample_results(
+        self, true_labels: Optional[np.ndarray] = None, label_map: Optional[dict[int, str]] = None
+    ) -> pd.DataFrame:
+        """
+        Returns a DataFrame with true labels (if provided) and predicted cluster
+        labels for each sample, with text labels if a map is provided.
+        """
+        results_df = pd.DataFrame()
+        if true_labels is not None:
+            results_df["true_label"] = true_labels
+            if label_map:
+                results_df["true_label"] = results_df["true_label"].map(lambda x: label_map.get(x))
+
+        results_df["predicted_label"] = self.predicted_labels
+        if label_map:
+            # Map predicted labels, handling noise (-1) separately
+            predicted_map = {k: v for k, v in label_map.items() if k != -1}
+            results_df["predicted_label"] = (
+                results_df["predicted_label"].map(lambda x: predicted_map.get(x)).fillna("Noise")
+            )
+
+        return results_df
+
+    def evaluate_against_truth(self, true_labels: np.ndarray) -> pd.DataFrame:
         """
         Calculates external clustering validation metrics against ground truth labels.
         """
-        ari = adjusted_rand_score(true_labels, self.predicted_labels)
-        nmi = normalized_mutual_info_score(true_labels, self.predicted_labels)
-        return {"ARI": ari, "NMI": nmi}
+        # Filter out noise from predictions for a fair comparison
+        mask = self.predicted_labels != -1
+        if not np.any(mask):
+            return pd.DataFrame([{"ARI": 0, "NMI": 0, "V-Measure": 0, "Cluster_Purity": 0}])
 
-    def evaluate_cluster_structure(self) -> dict[str, float]:
+        filtered_true = true_labels[mask]
+        filtered_pred = self.predicted_labels[mask]
+
[…]

[…]
