### 2025-09-19 ChristopherMarais/IBBI#35 @ChristopherMarais
We are looking into expanding the use of the package to Apple Silicon as well. No updates for this yet.
