### 2025-09-19 ba0267a
updated paper
