+++ README.md
@@ -10,24 +10,45 @@ This package is designed to support entomological research by automating the lab
 
 ### Motivation
 
-The ability to accurately identify bark and ambrosia beetles is critical for forest health and pest management. However, traditional methods face significant challenges:
+The ability to accurately detect and identify bark and ambrosia beetles is critical for forest health and pest management. However, traditional methods face significant challenges:
 
 * **They are slow and time-consuming.**
 * **They require highly specialized expertise.**
 * **They create a bottleneck for large-scale research.**
 
-The IBBI package provides a powerful, modern solution to overcome these obstacles by using pre-trained, open-source models to automate detection and classification from images, lowering the barrier to entry for researchers.
+The IBBI package provides a powerful, modern solution to overcome these obstacles by making available pre-trained, open-source models to automate detection and classification from images, lowering the barrier to entry for researchers. The IBBI package also makes other advanced techniques available to the user such as general zero-shot detection, general feature extraction, model evaluation, and model explainability.
 
 ### Key Features
 
-* **Simple API:** Access powerful detection and classification models with a single function call: `ibbi.create_model()`.
-* **Multiple Model Types:**
-    * **Single-Class Detection:** Detect the presence of *any* beetle in an image.
-    * **Multi-Class Species Detection:** Identify the species of a beetle from an image.
-    * **Zero-Shot Detection:** Detect objects using a text prompt (e.g., "insect"), without prior training on that specific class.
-* **Pre-trained Models:** Leverages pre-trained models hosted on the Hugging Face Hub for immediate use.
-* **Model Explainability:** Understand model predictions using SHAP (SHapley Additive exPlanations) to visualize which parts of an image contribute to the identification.
-* **Extensible:** Designed to easily incorporate new model architectures in the future.
+* **Model Access:**
+<br>Access powerful models with a single function call `ibbi.create_model()`. The following types of models are available:
+
+  * **Single-Class Bark Beetle Detection:** Detect the presence of *any* bark beetle in an image. These models have been trained with the task of single-class object detection on Bark and ambrosia beetle specific data. These models do not classify the species of beetle, only its presence and location in an image.
+
+  * **Multi-Class Species Detection:** Identify the species of a beetle from an image. These models have been trained with the task of multi-class object detection on Bark and ambrosia beetle specific data. These models classify the species of beetle, as well as its presence and location in an image.
+
+  * **General Zero-Shot Detection:** Detect objects using a text prompt (e.g., "insect"), without prior training on bark and ambrosia beetle specific data (e.g., GroundingDINO, YoloWorld). These models have been trained on large, diverse datasets and can generalize to detect a wide range of objects based on textual descriptions.
+
+  * **General Pre-trained Feature Extraction:** Extract feature embeddings from images using pre-trained models without additional training for specifically identifying bark and ambrosia beetles (e.g., DINOv3, EVA-02).
+
+* **Model Evaluation:**
+<br>This package includes a small set of ~2 000 images of 63 species for benchmarking model performance using a simple call to the `ibbi.Evaluator()` wrapper. The evaluation wrapper makes computing of the following types of metrics easy:
+
[…]
