### 2025-09-20 c1c543e
update README with latest code examples
