+++ README.md
@@ -129,16 +129,13 @@ Using `ibbi` is straightforward. Load a model and immediately use it for inferen
 import ibbi
 
 # --- List Available Models ---
-# The `ibbi` package provides a handy function to see all available models.
-print("Available models in the `ibbi` package:")
 ibbi.list_models()
 
 # --- Create a Model ---
-# Let's create a multi-class object detection model for classifying bark beetles.
 classifier = ibbi.create_model(model_name="species_classifier", pretrained=True)
-# For more specific models, refer to the model list
-# and replace model_name with your choice.
-# e.g., "yolov12x_bb_multi_class_detect_model"
+
+# --- Perform Inference ---
+results = classifier.predict("path/to/your/image.jpg")
 
 ```
 
@@ -151,24 +148,63 @@ For more detailed demonstrations, please see the example notebooks located in th
 To see a list of available models and their performance metrics directly from Python, run:
 
 ```python
-
+ibbi.list_models()
 ```
+Replace `model_name` with one of the available model names to use those models.
 
 **Model Summary Table**
 
+The most detailed version of the table can also be found [`here`](src/ibbi/data/ibbi_model_summary.csv).
 
-A detailed list can also be found in the [`src/ibbi/data/ibbi_model_summary.csv`](src/ibbi/data/ibbi_model_summary.csv) file.
+| **Model Name**                                      | **Tasks**                                                              | **Pretrained Weights Repository**                                                 | **Paper**                            | **Embedding vector shape** | **# of images fine-tuned** | **mAP@[.5:.95]** | **F1-score (Macro)** | **Silhouette score** | **Cluster Purity** |
+|-----------------------------------------------------|------------------------------------------------------------------------|-----------------------------------------------------------------------------------|--------------------------------------|----------------------------|----------------------------|------------------|----------------------|----------------------|--------------------|
+| **grounding_dino_detect_model**                     | Zero-shot Object Detection (Prompt: 'bark beetle'), Feature extraction | https://huggingface.co/IDEA-Research/grounding-dino-base                          | https://arxiv.org/abs/2303.05499     | (1, 256)                   | 0                          | 0.673            | 0.969                | -0.239               | 0.997              |
+| **yoloworldv2_bb_detect_model**                     | Zero-shot Object Detection (Prompt: 'bark beetle'), Feature extraction | https://github.com/ultralytics/assets/releases/download/v8.3.0/yolov8x-worldv2.pt | https://arxiv.org/pdf/2401.17270v2   | ( , 640)                   | 0                          | 0.004            | 0.007                | 0.686                | 0.577              |
+| **yolov8x_bb_detect_model**                         | Single-class Object Detection, Feature extraction                      | https://huggingface.co/IBBI-bio/ibbi_yolov8_od                                    | https://arxiv.org/abs/2408.15857     | ( , 640)                   | 46 781                     | 0.985            | 0.988                | 0.558                | 0.554              |
+| **yolov9e_bb_detect_model**                         | Single-class Object Detection, Feature extraction                      | https://huggingface.co/IBBI-bio/ibbi_yolov9_od                                    | https://arxiv.org/abs/2402.13616     | ( , 512)                   | 46 781                     | 0.987            | 0.989                | 0.338                | 0.568              |
[…]
