+++ docs/usage.md
@@ -1,178 +1,303 @@
 # Usage Guide
 
-This guide will walk you through the primary functionalities of the `ibbi` package, from installation to prediction and feature extraction.
+This guide provides an in-depth walkthrough of the `ibbi` package's functionalities. We will cover everything from installation and basic setup to advanced applications like model evaluation and explainability, complete with detailed code examples and explanations to help you integrate `ibbi` into your research workflow.
 
 ## Installation
 
-First, ensure you have PyTorch installed according to the official instructions at [pytorch.org](https://pytorch.org/get-started/locally/). Then, install `ibbi` from PyPI:
+Getting `ibbi` set up on your system is a two-step process. Since `ibbi` relies on PyTorch for its deep learning capabilities, it's crucial to install it first to ensure compatibility with your hardware, especially if you have a GPU.
+
+**1. Install PyTorch**
+
+For optimal performance, particularly with GPU acceleration, it is essential to install the correct version of PyTorch for your system (Windows/Mac/Linux) and CUDA version.
+Please follow the official, system-specific instructions at [pytorch.org](https://pytorch.org/get-started/locally/). This will ensure you leverage the full power of your hardware.
+
+**2. Install IBBI**
+
+Once PyTorch is installed, you can install `ibbi` from the Python Package Index (PyPI) using pip. It is highly recommended to do this within a virtual environment to avoid conflicts with other projects.
+
+<details>
+<summary>🐍 Conda Installation</summary>
 
 ```bash
+# 1. Create and activate a new conda environment
+conda create -n ibbi_env python=3.11 -y
+conda activate ibbi_env
+
+# 2. Install PyTorch first by following the official instructions
+# for your specific system at https://pytorch.org/get-started/locally/
+
+# 3. Install the ibbi package using pip
 pip install ibbi
 ```
+</details>
 
----
+<details>
+<summary>🟨 Pixi Installation</summary>
+
+```bash
+# 1. Initialize a pixi project in your directory (optional)
+pixi init
+
+# 2. Install PyTorch first by following the official instructions
+# for your specific system at https://pytorch.org/get-started/locally/
+
+# 3. Add ibbi to your project. Pixi will automatically handle the
+# creation of the environment and resolve all dependencies, including PyTorch.
+pixi add ibbi
+
+# 4. Run commands or scripts within the managed environment
+# For example, to start an interactive python session:
+pixi run python
+```
+</details>
+
+<br>For developers who want to contribute or use the latest, unreleased features, you can install the package directly from the source on GitHub:
+
+```bash
+pip install git+https://github.com/ChristopherMarais/IBBI.git
+```
 
 ## Core Functions
 
-The `ibbi` package is designed for simplicity. Here are the core functions you will use.
+The `ibbi` package is designed around a simple and intuitive API. The following core functions are your primary entry points for exploring, loading, and using the models.
 
 ### Listing Available Models
 
-To see all available models and their performance metrics, use `ibbi.list_models()`. This is useful for choosing the best model for your needs.
+Before you can use a model, you need to know what's available. The `ibbi.list_models()` function provides a comprehensive overview of all registered models, their tasks, and key performance metrics. This is the best first step to help you select the most appropriate model for your specific needs.
 
 ```python
 import ibbi
 
-# Get the model list as a pandas DataFrame
+# Get the model list as a pandas DataFrame for easy viewing and filtering
 models_df = ibbi.list_models(as_df=True)
-print(models_df)
+
+# Display the full table to see all available models and their metrics
+print("--- Full Model Summary ---")
+print(models_df.to_string())
+
[…]
