### 2025-09-22 3458a3d
updated paper and docstrings
