+++ paper.md
@@ -7,7 +7,7 @@ tags:
   - deep learning
   - entomology
   - object detection
-  - classification
+  - object classification
   - forest health
 authors:
   - name: Christopher Marais
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 # Summary
 
-The Intelligent Bark Beetle Identifier (IBBI) is an open-source Python package that provides a simple, unified interface for the automated detection and classification of bark and ambrosia beetles (Coleoptera: Curculionidae: Scolytinae and Platypodinae) from images. This package addresses a critical need in forest health and biodiversity research by leveraging multiple state-of-the-art computer vision models to streamline a traditionally labor-intensive and expert-dependent task. The target audience for `ibbi` includes entomologists, ecologists, forest health professionals, and biodiversity researchers who require high-throughput identification of beetle specimens from field or lab images. By providing programmatic access to a suite of pre-trained models for object detection and species classification, `ibbi` empowers researchers to accelerate data analysis for ecological studies, invasive species monitoring, and biodiversity assessments. The core functionality is accessible through a single factory function, `create_model()`, which simplifies model loading and inference, making deep learning techniques accessible to users without requiring specialized expertise in computer vision.
+The Intelligent Bark Beetle Identifier (IBBI) is an open-source Python package that provides a simple, unified interface for the automated detection and classification of bark and ambrosia beetles (Coleoptera: Curculionidae: Scolytinae and Platypodinae) from images. This package addresses a critical need in forest health and biodiversity research by leveraging multiple state-of-the-art computer vision models to streamline a traditionally labor-intensive and expert-dependent task. The target audience for `ibbi` includes entomologists, ecologists, forest health professionals, and biodiversity researchers who require high-throughput identification of beetle specimens from field or lab images. By providing programmatic access to a suite of pre-trained models for object-detection and species classification, `ibbi` empowers researchers to accelerate data analysis for ecological studies, invasive species monitoring, and biodiversity assessments. The core functionality is accessible through a single factory function, `create_model()`, which simplifies model loading and inference, making deep learning techniques accessible to users without requiring specialized expertise in computer vision.
 
 # Statement of Need
 
@@ -41,36 +41,47 @@ The accurate and timely identification of bark and ambrosia beetle species is fu
 
 While computer vision offers a promising solution, a critical gap exists in the available software ecosystem. There is a lack of specialized, accessible, and benchmarked tools for bark beetle identification. `ibbi` fills this gap by providing a robust, easy-to-install Python package that:
 1.  Focuses specifically on the taxonomically challenging and economically critical bark and ambrosia beetle groups.
-2.  Offers a range of pre-trained, benchmarked models for different tasks (object detection and classification, zero-shot object detection).
+2.  Offers a range of pre-trained, benchmarked models for different tasks (object-detection and object-classification, zero-shot object-detection).
 3.  Standardizes model access through a simple, high-level API, lowering the barrier to entry for researchers.
 
 By consolidating these functionalities into a single package, `ibbi` provides a foundational tool that enables reproducible, high-throughput research that was previously infeasible.
 
 # State of the Field
 
[…]
