+++ src/ibbi/explain/lime.py
@@ -22,14 +22,34 @@ from ..models.zero_shot import GroundingDINOModel
 
 
 def _prediction_wrapper(model: ModelType, text_prompt: Optional[str] = None) -> Callable:
-    """
-    Creates a highly efficient, batched prediction function for LIME.
-    This version processes the entire batch of images as a single tensor.
+    """Creates a highly efficient, batched prediction function for LIME.
+
+    This function acts as a factory, returning a `predict` function that is compatible
+    with LIME's expectation of processing a numpy array of perturbed images. It handles
+    different model types and ensures predictions are returned in the required format
+    (a numpy array of class probabilities).
+
+    Args:
+        model (ModelType): The instantiated `ibbi` model to be explained.
+        text_prompt (Optional[str], optional): A text prompt required for zero-shot models
+                                               like GroundingDINO. Defaults to None.
+
+    Returns:
+        Callable: A `predict` function that takes a batch of images as a numpy array
+                  and returns a numpy array of prediction probabilities.
     """
 
     def predict(image_array: np.ndarray, verbose: bool = False) -> np.ndarray:
-        """
-        Batched prediction function for LIME.
+        """Batched prediction function for LIME.
+
+        Args:
+            image_array (np.ndarray): A numpy array representing a batch of images,
+                                      with shape (num_samples, height, width, channels).
+            verbose (bool, optional): Whether to print prediction details. Defaults to False.
+
+        Returns:
+            np.ndarray: A 2D numpy array of shape (num_samples, num_classes) containing the
+                        prediction probabilities for each image and class.
         """
         if image_array.ndim == 3:
             image_array = np.expand_dims(image_array, 0)
@@ -84,9 +104,28 @@ def explain_with_lime(
     top_labels: int = 5,
     num_features: int = 100000,
 ) -> tuple[lime_image.ImageExplanation, Image.Image]:
-    """
-    Generates LIME explanations for a single image using batched predictions
-    and a faster segmentation algorithm.
+    """Generates LIME explanations for a single image.
+
+    This function uses the LIME (Local Interpretable Model-agnostic Explanations) algorithm
+    to explain a model's prediction on a single image. It identifies which parts (superpixels)
+    of the image were most influential in the model's decision-making process.
+
+    Args:
+        model (ModelType): The instantiated `ibbi` model to explain.
+        image (Image.Image): The PIL Image to be explained.
+        text_prompt (Optional[str], optional): A text prompt required for zero-shot models
+                                               like GroundingDINO. Defaults to None.
+        image_size (tuple[int, int], optional): The size to which the image will be resized for the explanation.
+                                                Defaults to (640, 640).
+        batch_size (int, optional): The number of perturbed images to process in a single batch. Defaults to 50.
+        num_samples (int, optional): The number of perturbed images to generate for the LIME algorithm. Defaults to 1000.
+        top_labels (int, optional): The number of top predicted classes to generate explanations for. Defaults to 5.
+        num_features (int, optional): The maximum number of superpixels to include in the explanation. Defaults to 100000.
+
+    Returns:
+        tuple[lime_image.ImageExplanation, Image.Image]: A tuple containing:
+            - The LIME explanation object from the `lime` library.
+            - The original, unmodified PIL Image.
     """
     original_image = image
     image_to_explain = image.resize(image_size)
@@ -111,11 +150,18 @@ def explain_with_lime(
     return explanation, original_image
 
 
-def plot_lime_explanation(
-    explanation: lime_image.ImageExplanation, image: Image.Image, top_k: int = 1, alpha: float = 0.6
-) -> None:
-    """
-    Plots a detailed LIME explanation with a red-to-green overlay.
[…]

[…]
