### 2025-09-22 fc49cad
Fix/remove conflict file (#37)
* update zero shot models to include the same functionality as other models

* yoloworld extract pred, confidence fix

* add cluster mtreics to summary table

* new package layout

* package layout: renamed modules and classes for usability

* added new example notebooks for new layout

* added new evaluation metrics

* updated distance matrix

* updated paper

* updated README

* update README with latest code examples

* updated docs md files

* updated paper and docstrings

* updated docs layout
