+++ src/ibbi/models/zero_shot.py
@@ -1,0 +1,287 @@
+# src/ibbi/models/zero_shot.py
+
+"""
+This module provides models for zero-shot object detection. These models are capable of
+detecting objects in images based on arbitrary text prompts, without being explicitly
+trained on a predefined set of classes. This makes them highly flexible for a wide
+range of detection tasks.
+
+The module includes two primary wrapper classes for different zero-shot architectures:
+- `GroundingDINOModel`: For the GroundingDINO model, which excels at open-set object detection.
+- `YOLOWorldModel`: For the YOLOWorld model, which extends the YOLO architecture with zero-shot capabilities.
+
+Additionally, it provides factory functions, decorated with `@register_model`, to easily
+instantiate these models with pretrained weights.
+"""
+
+from io import BytesIO
+from typing import Union
+
+import numpy as np
+import requests
+import torch
+from PIL import Image
+from transformers import AutoModelForZeroShotObjectDetection, AutoProcessor
+from ultralytics import YOLOWorld
+
+from ._registry import register_model
+
+
+class GroundingDINOModel:
+    """A wrapper class for the GroundingDINO zero-shot object detection model.
+
+    This class provides a standardized interface for using the GroundingDINO model for
+    detecting objects in an image based on a text prompt. It handles model and processor
+    loading from the Hugging Face Hub, device placement, and provides methods for both
+    prediction and feature extraction.
+
+    Args:
+        model_id (str, optional): The model identifier from the Hugging Face Hub.
+                                  Defaults to "IDEA-Research/grounding-dino-base".
+    """
+
+    def __init__(self, model_id: str = "IDEA-Research/grounding-dino-base"):
+        """Initializes the GroundingDINOModel.
+
+        Args:
+            model_id (str): The Hugging Face Hub model identifier for the GroundingDINO model.
+        """
+        self.processor = AutoProcessor.from_pretrained(model_id)
+        self.model = AutoModelForZeroShotObjectDetection.from_pretrained(model_id)
+        self.device = "cuda" if torch.cuda.is_available() else "cpu"
+        self.model.to(self.device)
+        print(f"GroundingDINO model loaded on device: {self.device}")
+
+    def get_classes(self):
+        """This method is not applicable to zero-shot models.
+
+        Raises:
+            NotImplementedError: As a zero-shot model, classes are defined dynamically
+                                 at inference time via the 'text_prompt' argument.
+        """
+        raise NotImplementedError(
+            "The GroundingDINOModel is a zero-shot detection model and does not have a fixed set of classes. "
+            "Classes are defined dynamically at inference time via the 'text_prompt' argument in the 'predict' method."
+        )
+
+    def predict(
+        self,
+        image,
+        text_prompt: str,
+        box_threshold: float = 0.25,
+        text_threshold: float = 0.25,
+        verbose: bool = False,
+    ):
+        """Performs zero-shot object detection on an image given a text prompt.
+
+        Args:
+            image (Union[str, np.ndarray, Image.Image]): The input image. Can be a file path, URL,
+                                                         numpy array, or PIL Image object.
+            text_prompt (str): The text prompt describing the object(s) to detect.
+            box_threshold (float, optional): The confidence threshold for filtering bounding boxes.
+                                             Defaults to 0.25.
+            text_threshold (float, optional): The confidence threshold for filtering text labels.
+                                              Defaults to 0.25.
+            verbose (bool, optional): If True, prints detailed detection results. Defaults to False.
+
+        Returns:
+            dict: A dictionary containing the detection results with keys for 'scores',
+                  'text_labels', and 'boxes'.
[…]
