### 2025-09-22 3891424
fix dependencies and zero-shot output format
