+++ pyproject.toml
@@ -34,10 +34,8 @@ lime = "^0.2.0.1"
 scikit-bio = "^0.7.0"
 timm = "^1.0.19"
 umap-learn = "^0.5.9.post2"
-hdbscan = "^0.8.40"
 torch = { version = ">=2.1", source = "pytorch_cu128" }
 torchvision = { version = ">=0.16.0", source = "pytorch_cu128" }
-clip = {git = "https://github.com/openai/CLIP.git"}
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.11.10"
