### 2025-09-24 7a74ed5
refactor eval for object classification
