### 2025-09-24 b51a86e
fix yoloworld inference and embedding
