### 2025-09-24 470a9ac
Feature/classification update 2 (#39)
* spacing fixed

* update get_dataset() description

* refactor eval for object classification

* fix yoloworld inference and embedding

* made embedding eval more robust
