+++ paper.bib
@@ -156,3 +156,14 @@
   year={2021},
   publisher={Nature Publishing Group UK London}
 }
+
+@article{matheson2014inaturalist,
+  title={inaturalist},
+  author={Matheson, Christie A},
+  journal={Reference Reviews},
+  volume={28},
+  number={8},
+  pages={36--38},
+  year={2014},
+  publisher={Emerald Group Publishing Limited}
+}
