+++ src/ibbi/models/zero_shot.py
@@ -37,7 +37,7 @@ class GroundingDINOModel:
 
     Args:
         model_id (str, optional): The model identifier from the Hugging Face Hub.
-                                  Defaults to "IDEA-Research/grounding-dino-base".
+                                Defaults to "IDEA-Research/grounding-dino-base".
     """
 
     def __init__(self, model_id: str = "IDEA-Research/grounding-dino-base"):
@@ -50,45 +50,65 @@ class GroundingDINOModel:
         self.model = AutoModelForZeroShotObjectDetection.from_pretrained(model_id)
         self.device = "cuda" if torch.cuda.is_available() else "cpu"
         self.model.to(self.device)
+        self.classes: list[str] = []
         print(f"GroundingDINO model loaded on device: {self.device}")
 
-    def get_classes(self):
-        """This method is not applicable to zero-shot models.
+    def get_classes(self) -> list[str]:
+        """Returns the classes the model is currently set to detect.
 
-        Raises:
-            NotImplementedError: As a zero-shot model, classes are defined dynamically
-                                 at inference time via the 'text_prompt' argument.
+        For zero-shot models, this is determined by the last `text_prompt` used.
+
+        Returns:
+            list[str]: A list of the class names currently set for detection.
         """
-        raise NotImplementedError(
-            "The GroundingDINOModel is a zero-shot detection model and does not have a fixed set of classes. "
-            "Classes are defined dynamically at inference time via the 'text_prompt' argument in the 'predict' method."
-        )
+        return self.classes
+
+    def set_classes(self, classes: Union[list[str], str]):
+        """Sets the classes for the model to detect.
+
+        Args:
+            classes (Union[list[str], str]): A list of class names or a single string
+                                            with class names separated by " . ".
+        """
+        if isinstance(classes, str):
+            self.classes = [c.strip() for c in classes.split(" . ")]
+        else:
+            self.classes = classes
+        # print(f"GroundingDINO classes set to: {self.classes}")
 
     def predict(
         self,
         image,
-        text_prompt: str,
-        box_threshold: float = 0.25,
-        text_threshold: float = 0.25,
+        text_prompt: Optional[str] = None,
+        box_threshold: float = 0.05,
+        text_threshold: float = 0.05,
         verbose: bool = False,
     ):
         """Performs zero-shot object detection on an image given a text prompt.
 
         Args:
             image (Union[str, np.ndarray, Image.Image]): The input image. Can be a file path, URL,
-                                                         numpy array, or PIL Image object.
-            text_prompt (str): The text prompt describing the object(s) to detect.
+                                                        numpy array, or PIL Image object.
+            text_prompt (str, optional): The text prompt describing the object(s) to detect.
+                                        If provided, this will set the detection classes for the model.
             box_threshold (float, optional): The confidence threshold for filtering bounding boxes.
-                                             Defaults to 0.25.
+                                            Defaults to 0.05.
             text_threshold (float, optional): The confidence threshold for filtering text labels.
-                                              Defaults to 0.25.
+                                            Defaults to 0.05.
             verbose (bool, optional): If True, prints detailed detection results. Defaults to False.
 
         Returns:
             dict: A dictionary containing the detection results with keys for 'scores',
-                  'labels', and 'boxes'.
+                'labels', and 'boxes'.
         """
-        print(f"Running GroundingDINO detection for prompt: '{text_prompt}'...")
+        if text_prompt:
+            self.set_classes(text_prompt)
+
+        if not self.classes:
+            raise ValueError("No classes set for detection. Please provide a 'text_prompt' or call 'set_classes' first.")
+
[…]
