+++ pyproject.toml
@@ -36,6 +36,7 @@ timm = "^1.0.19"
 umap-learn = "^0.5.9.post2"
 torch = { version = ">=2.1", source = "pytorch_cu128" }
 torchvision = { version = ">=0.16.0", source = "pytorch_cu128" }
+tabulate = "^0.9.0"
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.11.10"

[…]
