### 2025-09-24 ChristopherMarais/IBBI#34 @ChristopherMarais
@PetervanLunteren

The benchmark table is under [Available Models](https://github.com/ChristopherMarais/IBBI#available-models).

I also give link to it showing it [here](https://github.com/ChristopherMarais/IBBI/blob/main/src/ibbi/data/ibbi_model_summary.csv).

This table will be continuously be updated. I specifically have a few updates that I will make to it this week.

I have also now updated incomplete citation. It must have slipped past me, thank you for the reminder!
