+++ notebooks/ibbi_ood_dataset/README.md
@@ -1,0 +1,238 @@
+---
+license: mit
+dataset_info:
+  features:
+  - name: image
+    dtype: image
+  - name: objects
+    sequence:
+    - name: bbox
+      sequence: float64
+    - name: category
+      dtype: string
+  splits:
+  - name: train
+    num_examples: 2418
+  # Note: num_bytes and download_size are updated automatically by the Hub
+  # after the upload is complete. These are placeholder estimates.
+  download_size: 6432490239
+  dataset_size: 7697416830
+configs:
+- config_name: default
+  data_files:
+  - split: train
+    path: data/train-*
+---
+
+# Dataset Card for IBBI Out-of-Distribution (OOD) Dataset
+
+## Dataset Summary
+
+This dataset contains **out-of-distribution (OOD)** images of bark and ambrosia beetles, intended for evaluating the robustness and generalization capabilities of models from the `ibbi` Python package. The **121 species** included in this dataset were not part of the original training or in-distribution test sets for the models in the `ibbi` package.
+
+This dataset is crucial for testing how well models can:
+- **Reject unknown classes** (for multi-class classifiers) by assigning low confidence scores.
+- **Generalize detection** to beetle species they have not been explicitly trained on (for single-class detectors).
+
+The dataset can be loaded directly using the `ibbi` package:
+```python
+import ibbi
+ood_dataset = ibbi.get_dataset(repo_id="IBBI-bio/ibbi_ood_dataset")
+````
+
+## Supported Tasks
+
+  - **Out-of-Distribution Detection**: The primary purpose is to evaluate how models respond to unseen species.
+  - **Object Detection**: Can be used to benchmark the generalization performance of beetle detectors.
+
+## Dataset Structure
+
+### Data Instances
+
+Each instance in the dataset consists of an image and a corresponding set of object annotations.
+
+**Example:**
+
+```json
+{
+  "image": "<PIL.Image.Image image>",
+  "objects": {
+    "bbox": [
+      [656.56, 1335.47, 80.53, 75.35]
+    ],
+    "category": [
+      "Dendroctonus_frontalis"
+    ]
+  }
+}
+```
+
+### Data Fields
+
+  * `image`: A PIL Image object of a beetle specimen.
+  * `objects`: A dictionary containing the annotation information for the image.
+      * `bbox`: A list of bounding boxes. Each box is in the format `[x_min, y_min, width, height]`.
+      * `category`: A list of string labels corresponding to the species in each bounding box.
+
+### Data Splits
+
+The dataset contains a single split, named `train` for compatibility with Hugging Face's default structure. This split serves as the official out-of-distribution evaluation set.
+
+  * **`train`**: 2,418 images with annotations.
+
+## Dataset Statistics
+
+| Metric                         | Value         |
+| ------------------------------ | ------------- |
+| Total Images with Annotations  | 2,418         |
+| Unique Species                 | 121           |
+| Total Bounding Box Annotations | 12,884        |
+| Average Bboxes per Image       | 5.33          |
+| Average Image Dimensions (WxH) | 2146 x 1702   |
+
+### Species Distribution
+
+| Species                             | Annotation Count |
+| ----------------------------------- | ---------------- |
+| Pseudopityophthorus\_minutissimus    | 1220             |
+| Dactylotrypes\_longicollis           | 1090             |
+| Carphoborus\_bifurcus                | 1077             |
+| Scolytus\_rugulosus                  | 867              |
+| Hypothenemus\_seriatus               | 840              |
+| Anisandrus\_maiche                   | 643              |
+| Scolytus\_carpini                    | 620              |
+| Cryphalus\_mangiferae                | 612              |
[…]
