### 2025-10-01 d9a8206
Feature/v0.2.2 edits (#41)
* fixed mantel's R and removed V-measure

* added genus classification

* OOD evaluation

* added codemeta.json file
