+++ src/ibbi/evaluate/embeddings.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, Optional, cast
 
 import numpy as np
 import pandas as pd
-from scipy.spatial.distance import pdist, squareform
+import torch
 from sklearn.cluster import HDBSCAN
 from sklearn.metrics import (
     adjusted_rand_score,
@@ -13,7 +13,6 @@ from sklearn.metrics import (
     davies_bouldin_score,
     normalized_mutual_info_score,
     silhouette_score,
-    v_measure_score,
 )
 from sklearn.preprocessing import LabelEncoder
 
@@ -186,18 +185,18 @@ class EmbeddingEvaluator:
             true_labels (np.ndarray): An array of ground truth labels for each sample.
 
         Returns:
-            pd.DataFrame: A DataFrame containing external validation metrics like ARI, NMI, V-Measure, and Cluster Purity.
+            pd.DataFrame: A DataFrame containing external validation metrics like ARI, NMI, and Cluster Purity.
         """
         # Filter out noise from predictions for a fair comparison
         mask = self.predicted_labels != -1
         if not np.any(mask):
-            return pd.DataFrame([{"ARI": 0, "NMI": 0, "V-Measure": 0, "Cluster_Purity": 0}])
+            return pd.DataFrame([{"ARI": 0, "NMI": 0, "Cluster_Purity": 0}])
 
         filtered_true = true_labels[mask]
         filtered_pred = self.predicted_labels[mask]
 
         if len(np.unique(filtered_true)) < 2 or len(np.unique(filtered_pred)) < 2:
-            return pd.DataFrame([{"ARI": 0, "NMI": 0, "V-Measure": 0, "Cluster_Purity": 0}])
+            return pd.DataFrame([{"ARI": 0, "NMI": 0, "Cluster_Purity": 0}])
 
         le_true = LabelEncoder().fit(filtered_true)
         true_labels_encoded = le_true.transform(filtered_true)
@@ -205,10 +204,9 @@ class EmbeddingEvaluator:
 
         ari = adjusted_rand_score(true_labels_encoded, predicted_labels_encoded)
         nmi = normalized_mutual_info_score(true_labels_encoded, predicted_labels_encoded)
-        v_measure = v_measure_score(true_labels_encoded, predicted_labels_encoded)
         purity = _cluster_purity(true_labels_encoded, predicted_labels_encoded)
 
-        metrics = {"ARI": ari, "NMI": nmi, "V-Measure": v_measure, "Cluster_Purity": purity}
+        metrics = {"ARI": ari, "NMI": nmi, "Cluster_Purity": purity}
 
         return pd.DataFrame([metrics])
 
@@ -242,12 +240,14 @@ class EmbeddingEvaluator:
     ) -> tuple[float, float, int, pd.DataFrame]:
         """Calculates Mantel correlation between embedding distances and an external distance matrix.
         The default is to use a distance matrix based on phylogenetic and taxonomic distance between species.
+        This version computes the average pairwise distances between all embeddings for each pair of species
+        and leverages a GPU if available.
 
         Args:
             true_labels (np.ndarray): An array of ground truth labels for each sample.
             label_map (Optional[dict[int, str]], optional): A dictionary to map integer labels to string names.
                                                         Defaults to None.
-            embedding_metric (str, optional): The distance metric to use for the embedding space.
+            embedding_metric (str, optional): The distance metric to use for the embedding space ('cosine' or 'euclidean').
                                             Defaults to "cosine".
             ext_dist_matrix_path (str, optional): The path to the external distance matrix file.
                                                 Defaults to "ibbi_species_distance_matrix.csv".
@@ -255,30 +255,56 @@ class EmbeddingEvaluator:
         Returns:
[…]
