+++ src/ibbi/evaluate/object_classification.py
@@ -10,7 +10,7 @@ ground truth objects before comparing their class labels.
 """
 
 from collections import defaultdict
-from typing import Any, Union
+from typing import Any, Optional, Union
 
 import numpy as np
 import pandas as pd
@@ -110,6 +110,7 @@ def object_classification_performance(
                           and a classification report.
                         - "sample_results": A pandas DataFrame with detailed information on each
                                             ground truth and predicted box for error analysis.
+                        - "object_table": A pandas DataFrame with detailed information on each matched object.
     """
     if iou_thresholds is None:
         iou_thresholds = np.arange(0.5, 1.0, 0.05)
@@ -163,6 +164,7 @@ def object_classification_performance(
     per_threshold_scores = {}
     per_iou_classification_metrics = {}
     aps_last_iou = {}
+    object_table = []
 
     # --- mAP Calculation Loop ---
     for iou_threshold in iou_thresholds:
@@ -192,6 +194,7 @@ def object_classification_performance(
                 gt_data = gt_by_image[pred["image_id"]]
                 best_iou = -1.0
                 best_gt_idx = -1
+                best_gt_box = None
 
                 for j, gt_box in enumerate(gt_data["boxes"]):
                     if gt_data["labels"][j] == class_id:
@@ -199,11 +202,22 @@ def object_classification_performance(
                         if iou > best_iou:
                             best_iou = iou
                             best_gt_idx = j
+                            best_gt_box = gt_box
 
                 if best_iou >= iou_threshold:
                     if not gt_data["used"][best_gt_idx]:
                         tp[i] = 1
                         gt_data["used"][best_gt_idx] = True
+                        object_table.append(
+                            {
+                                "image_id": pred["image_id"],
+                                "gt_bbox": best_gt_box,
+                                "pred_bbox": pred["box"],
+                                "iou": best_iou,
+                                "class_id": class_id,
+                                "score": pred["score"],
+                            }
+                        )
                     else:
                         fp[i] = 1
                 else:
@@ -289,6 +303,7 @@ def object_classification_performance(
         }
 
     final_map_averaged = np.mean(list(per_threshold_scores.values())) if per_threshold_scores else 0.0
+    object_df = pd.DataFrame(object_table)
 
     return {
         "mAP": final_map_averaged,
@@ -296,4 +311,191 @@ def object_classification_performance(
         "per_threshold_scores": per_threshold_scores,
         "per_iou_classification_metrics": per_iou_classification_metrics,
         "sample_results": pd.DataFrame(sample_results),
+        "object_table": object_df,
+    }
+
+
+def out_of_distribution_detection(
+    id_results: list[dict],
+    ood_results: list[dict],
+    id_dataset: list[dict],
+    ood_dataset: list[dict],
+    class_names: list[str],
+    thresholds: Optional[np.ndarray] = None,
+) -> dict[str, Any]:
+    """Performs a comprehensive out-of-distribution (OOD) analysis.
+
+    This function compares the model's behavior on in-distribution (ID) versus
+    out-of-distribution (OOD) data. It provides detailed statistics on confidence
+    scores, allowing for the determination of an optimal threshold for separating
+    known from unknown classes.
+
+    Args:
+        id_results (list[dict]): A list of prediction results from the model on the ID dataset.
+        ood_results (list[dict]): A list of prediction results from the model on the OOD dataset.
+        id_dataset (list[dict]): The in-distribution dataset.
+        ood_dataset (list[dict]): The out-of-distribution dataset.
[…]
