### 2025-10-02 5e471d0
batch processing fix for embeddings eval
