### 2025-10-02 a26ed36
Feature/v0.2.2 edits (#42)
* fixed mantel's R and removed V-measure

* added genus classification

* OOD evaluation

* added codemeta.json file

* batch processing fix for embeddings eval
