### 2025-10-03 2d15b6e
fix all models to run eval
