+++ src/ibbi/evaluate/object_classification.py
@@ -58,6 +58,7 @@ def object_classification_performance(
     gt_labels: list[int],
     gt_image_ids: list[Any],
     pred_results_with_probs: list[dict],
+    gt_label_names: list[str],
     iou_thresholds: Union[float, list[float], np.ndarray] | None = None,
     confidence_threshold: float = 0.5,
     average: str = "macro",
@@ -77,6 +78,7 @@ def object_classification_performance(
         gt_image_ids (list[Any]): A list of image identifiers for each ground truth box.
         pred_results_with_probs (list[dict]): A list of prediction results (one per image), where each
                                               result dictionary must contain 'full_results' with per-class probabilities.
+        gt_label_names (list[str]): A list of the original ground truth label names from the dataset.
         iou_thresholds (Union[float, list[float], np.ndarray], optional): The IoU threshold(s)
             for mAP and classification metric calculation. Defaults to np.arange(0.5, 1.0, 0.05).
         confidence_threshold (float, optional): The confidence score threshold below which
@@ -114,13 +116,14 @@ def object_classification_performance(
         full_results = res.get("full_results", [None] * len(res["boxes"]))
 
         for box, label, score, full_result in zip(res["boxes"], res["labels"], res["scores"], full_results):
-            # Only consider predictions above the confidence threshold and matching a known class name
-            if score >= confidence_threshold and label in name_to_idx:
+            # Only consider predictions above the confidence threshold
+            if score >= confidence_threshold:
                 # For mAP/Classification (needs flat lists of integer IDs)
-                pred_boxes.append(np.array(box).flatten())
-                pred_labels.append(name_to_idx[label])
-                pred_scores.append(score)
-                pred_image_ids.append(image_id)
+                if label in name_to_idx:
+                    pred_boxes.append(np.array(box).flatten())
+                    pred_labels.append(name_to_idx[label])
+                    pred_scores.append(score)
+                    pred_image_ids.append(image_id)
 
                 # For the object_level_performance table
                 image_to_preds[image_id].append({"box": box, "score": score, "label_name": label, "full_results": full_result})
@@ -133,8 +136,10 @@ def object_classification_performance(
 
     all_gt_labels = list(set(gt_labels))
     all_pred_labels = list(set(pred_labels))
-    class_lst = list(set(all_gt_labels) | set(all_pred_labels))
-    all_classes = sorted(class_lst)
+    class_lst_us = list(set(all_gt_labels) | set(all_pred_labels))
+    class_lst = sorted(class_lst_us)
+    all_classes_us = list(set(c for c in class_lst if c != -1))  # Exclude -1 from all_classes
+    all_classes = sorted(all_classes_us)
 
     # Group GT by image for faster lookups and 'used' tracking
     gt_by_image = defaultdict(lambda: {"boxes": [], "labels": [], "used": []})
@@ -249,12 +254,12 @@ def object_classification_performance(
         kappa = cohen_kappa_score(true_class_labels_for_iou, pred_class_labels_for_iou) if has_multiple_effective_classes else np.nan
         mcc = matthews_corrcoef(true_class_labels_for_iou, pred_class_labels_for_iou) if has_multiple_effective_classes else np.nan
 
-        extended_classes_lst = list(set(all_classes) | {-1})
-        extended_classes = sorted(extended_classes_lst)
+        extended_classes_us = list(set(all_classes) | {-1})
+        extended_classes = sorted(extended_classes_us)
         extended_target_names = [idx_to_name.get(c, "No_Match") for c in extended_classes]
 
-        class_subset_for_metrics_lst = list(set(true_class_labels_for_iou) - {-1})
-        class_subset_for_metrics = sorted(class_subset_for_metrics_lst)
[…]
