### 2025-10-03 de8dcb7
fixed yoloworld eval
