### 2025-10-04 d8e1a23
fixed groundingdino eval
