### 2025-10-04 58a0e0d
Feature/v0.2.2 edits (#43)
* fixed mantel's R and removed V-measure

* added genus classification

* attempt 1 for OOD evaluation

* added codemeta.json file

* batch processing fix for embeddings eval

* fix all models to run eval

* fix output tables for eval

* fixed yoloworld eval

* fixed groundingdino eval
