+++ src/ibbi/evaluate/object_classification.py
@@ -1,4 +1,3 @@
-# src/ibbi/evaluate/object_classification.py
 """
 Provides a unified function for evaluating object detection and classification performance.
 
@@ -58,88 +57,96 @@ def object_classification_performance(
     gt_boxes: np.ndarray,
     gt_labels: list[int],
     gt_image_ids: list[Any],
-    pred_boxes: np.ndarray,
-    pred_labels: list[int],
-    pred_scores: list[float],
-    pred_image_ids: list[Any],
+    pred_results_with_probs: list[dict],
+    gt_label_names: list[str],
     iou_thresholds: Union[float, list[float], np.ndarray] | None = None,
-    confidence_threshold: float = 0.5,
+    confidence_threshold: float = 0.1,
     average: str = "macro",
     zero_division: Union[str, int, float] = np.nan,
+    model_classes: Optional[list[str]] = None,
+    idx_to_name: Optional[dict[int, str]] = None,
 ) -> dict[str, Any]:
     """Calculates a comprehensive suite of object detection and classification metrics.
 
-    This function provides a holistic evaluation of an object detection model's performance.
-    It computes the mean Average Precision (mAP) over a range of Intersection over Union (IoU)
-    thresholds to assess localization accuracy. Simultaneously, it calculates a full suite of
-    classification metrics (accuracy, precision, recall, F1-score, etc.) for each IoU threshold by
-    matching predicted objects to ground truth objects.
-
-    This unified approach allows for a detailed breakdown of a model's ability to both find
-    and correctly identify objects at various levels of localization stringency.
+    This function performs a holistic evaluation, combining mAP and standard classification
+    metrics. It also generates a detailed object-level table mapping ground truth objects
+    to their best-matching predictions and full per-class confidence scores.
 
     Args:
-        gt_boxes (np.ndarray): A NumPy array of ground truth bounding boxes, with each box
-                               in [x1, y1, x2, y2] format. Shape: (N, 4).
-        gt_labels (list[int]): A list of integer labels corresponding to each ground truth box. Length: N.
-        gt_image_ids (list[Any]): A list of image identifiers for each ground truth box, used to
-                                  group boxes by image. Length: N.
-        pred_boxes (np.ndarray): A NumPy array of predicted bounding boxes in [x1, y1, x2, y2] format. Shape: (M, 4).
-        pred_labels (list[int]): A list of predicted integer labels for each predicted box. Length: M.
-        pred_scores (list[float]): A list of confidence scores for each predicted box. Length: M.
-        pred_image_ids (list[Any]): A list of image identifiers for each predicted box. Length: M.
+        gt_boxes (np.ndarray): A NumPy array of ground truth bounding boxes, in [x1, y1, x2, y2] format.
+        gt_labels (list[int]): A list of integer labels corresponding to each ground truth box.
+        gt_image_ids (list[Any]): A list of image identifiers for each ground truth box.
+        pred_results_with_probs (list[dict]): A list of prediction results (one per image), where each
+                                              result dictionary must contain 'full_results' with per-class probabilities.
+        gt_label_names (list[str]): A list of the original ground truth label names from the dataset.
         iou_thresholds (Union[float, list[float], np.ndarray], optional): The IoU threshold(s)
-            for mAP and classification metric calculation. Can be a single float or a list/array of floats.
-            Defaults to np.arange(0.5, 1.0, 0.05).
+            for mAP and classification metric calculation. Defaults to np.arange(0.5, 1.0, 0.05).
         confidence_threshold (float, optional): The confidence score threshold below which
-            predictions are ignored. Defaults to 0.5.
[…]
