+++ src/ibbi/evaluate/__init__.py
@@ -1,4 +1,5 @@
 # src/ibbi/evaluate/__init__.py
+
 """
 Provides the high-level Evaluator class for comprehensive model assessment.
 
@@ -15,10 +16,10 @@ import numpy as np
 from tqdm import tqdm
 
 from ..models import ModelType
-from ..models.feature_extractors import HuggingFaceFeatureExtractor
+from ..models.feature_extractors import HuggingFaceFeatureExtractor, UntrainedFeatureExtractor
 from ..models.zero_shot import GroundingDINOModel, YOLOWorldModel
 from .embeddings import EmbeddingEvaluator
-from .object_classification import object_classification_performance, out_of_distribution_detection
+from .object_classification import object_classification_performance
 
 
 class Evaluator:
@@ -78,87 +79,100 @@ class Evaluator:
 
         Returns:
             dict: A dictionary containing a comprehensive set of object detection and
-                  classification metrics, including per-IoU threshold classification performance.
-                  Returns an empty dictionary if the model is not suitable for the task or if
-                  the dataset is improperly formatted.
+                  classification metrics, including per-iou threshold classification performance,
+                  and a detailed object-level performance table.
         """
         if predict_kwargs is None:
             predict_kwargs = {}
 
+        # Set classes for GroundingDINO before evaluation
+        if isinstance(self.model, GroundingDINOModel):
+            if "text_prompt" in predict_kwargs:
+                self.model.set_classes(predict_kwargs["text_prompt"])
+
         print("Running object classification evaluation...")
 
-        if isinstance(self.model, HuggingFaceFeatureExtractor):
+        if isinstance(self.model, (HuggingFaceFeatureExtractor, UntrainedFeatureExtractor)):
             print("Warning: Object classification evaluation is not supported for pure feature extractors.")
             return {}
 
         if isinstance(self.model, (GroundingDINOModel, YOLOWorldModel)):
-            if "text_prompt" in predict_kwargs:
-                self.model.set_classes(predict_kwargs["text_prompt"])
-            elif not self.model.get_classes():
+            if "text_prompt" not in predict_kwargs and not self.model.get_classes():
                 print("Warning: Zero-shot model has no classes set. Please provide a 'text_prompt' in 'predict_kwargs'.")
                 return {}
 
-        if not hasattr(self.model, "get_classes") or not callable(self.model.get_classes):
-            print("Warning: Model does not have a 'get_classes' method for class mapping. Skipping evaluation.")
-            return {}
-
-        raw_model_classes = self.model.get_classes()
-        if isinstance(raw_model_classes, dict):
-            model_classes: list[str] = list(raw_model_classes.values())
-        else:
-            model_classes: list[str] = raw_model_classes
-        class_name_to_idx = {v: k for k, v in enumerate(model_classes)}
-        idx_to_name = dict(enumerate(model_classes))
-
-        gt_boxes, gt_labels, gt_image_ids = [], [], []
-        pred_boxes, pred_labels, pred_scores, pred_image_ids = [], [], [], []
+        gt_boxes, gt_labels, gt_image_ids, gt_label_names = [], [], [], []
+        pred_results_with_probs = []  # Full prediction result per image
+        # Initialize model_classes before the loop.
+        model_classes: list[str] = []
+        if isinstance(self.model, (GroundingDINOModel)):
+            if hasattr(self.model, "get_classes") and callable(self.model.get_classes):
+                raw_model_classes = self.model.get_classes()
+                if isinstance(raw_model_classes, dict):
+                    model_classes = list(raw_model_classes.values())
+                else:
+                    model_classes = raw_model_classes
+        class_name_to_idx: dict[str, int] = {}
+        idx_to_name: dict[int, str] = {}
 
         print("Extracting ground truth and making predictions...")
[…]

[…]
