### 2025-10-21 415d310
embedding adjustment for feature extractors
