### 2025-10-24 #8796 @goldingn
Sorry this is dragging on, @ChristopherMarais. I've contacted Martí via email, just in case there's an issue with GitHub messages getting through. I'll check back in mid next week. If we haven't heard back from Martí, I'll aim to do the second review myself.
