### 2025-11-07 #8796 @martibosch
Hello! my apologies for the delay, here is my review:

## first impressions

First of all I believe that this is a great package that tries to make it as easy as possible to use deep learning models for common tasks (without requiring proficiency in the torch stack). As a non-expert in deep learning, I particularly appreciate this kind of packages (e.g., like [deepforest](https://github.com/weecology/DeepForest)).

## typos and syntactic corrections

### manuscript

The manuscript and documentation are very well written and cover the review of related tools and the gap filled (statement of need) by the library. From the manuscript I would only suggest changing from Goodwin et al. (2021) to (Goodwin et al., 2021) in line 53 and idem for Sun et al. in line 57.

### documentation

I believe that the images of the usage guide are broken. This may be linked to the fact that the images are huge - I would suggest generating smaller image renditions for the user guide. Additionally, I would also suggest adding the code to generate the output images with predictions within the usage guide.

## content corrections

### user guide

Since I assume that the a large part of the target audience have little expertise in deep learning, the main thing that I have missed is a workflow to evaluate a custom image dataset and perhaps to fine-tune a pre-trained model too. I suppose (maybe I am wrong) that this package may be of interest to an ecologist with a collection of images on a folder with some directory organization or some csv file indexing the file names and attributes (and perhaps annotations too). From this perspective, I believe that an example on how to convert this to a Python object that can be passed as the dataset argument of, e.g., `ibbi.Evaluator.classification` would be helpful.

Also trying to run the ["Model evaluation"](https://gcmarais.com/IBBI/usage/#model-evaluation) snippet crashed the Jupyter kernel not only locally (on my CPU) but also on a T4 GPU in Collab. I further tried to slice the data, e.g., `data[:10]` but then I got a TypeError (see [the Collab notebook in question](https://colab.research.google.com/drive/13LS2dd6XIUMUG9Svn-aQTmXopgBC6rYw?usp=sharing) for the traceback).

### install and set up instructions

I feel like users would benefit from better recommendations on how (hardware-wise) to use ibbi. I tried to set up ibbi locally (on CPU) using pixi and I had to add the `--pypi` flag as in `pixi add --pypi ibbi`, otherwise no candidates are found (since ibbi is not yet in conda-forge). Besides the 10-20 GB of disk used by the installation and model/data downloads, it crashed when running the model evaluation (as mentioned above), so I wonder which are the memory requirements to run inference on such large images. Additionally, adding a link to a Collab notebook (or any other free GPU enabled kernel - I by no means mean to advertise Google products)  may be a good way to provide potential users an easy (0-setup) way to get started ant test the features of ibbi.

## further suggestions (beyond the scope of the JOSS review)

- I would strongly suggest to add tests and a CI pipeline, from my experience I can tell that this is essential to ensure that everything works as new features are added
- A very opinionated suggestion strongly biased by my experience: I am a geospatial data scientist who occasionally uses deep learning for computer vision, however the deep learning parts are in most cases only a small part of my overall pipeline. I believe that this may also be the case of a large portion of your target audience. Under these circumstances, I find it impractical to move all my computational pipeline to the cloud just to run a couple of its steps on a GPU, which is why I recently became interested in serverless computing so that I can keep everything local except those steps. At the risk of shameless self promotion, consider the [deepforest-modal-app](https://github.com/martibosch/deepforest-modal-app) that I created to use DeepForest in a serverless manner. I feel that ibbi could be another great case for this approach.
