### 2025-11-11 c077575
docs fix image renbdering
