### 2025-11-11 40f7109
added custom data evaluation notebook example
