### 2025-11-11 493e4e3
added isntructions for subselecting testing dataset
