+++ README.md
@@ -156,26 +156,29 @@ Replace `model_name` with one of the available model names to use those models.
 
 The most detailed version of the table can also be found [`here`](src/ibbi/data/ibbi_model_summary.csv).
 
-| **Model Name**                                      | **Tasks**                                                              | **Pretrained Weights Repository**                                                 | **Paper**                            | **Embedding vector shape** | **# of images fine-tuned** | **mAP@[.5:.95]** | **F1-score (Macro)** | **Silhouette score** | **Cluster Purity** |
-|-----------------------------------------------------|------------------------------------------------------------------------|-----------------------------------------------------------------------------------|--------------------------------------|----------------------------|----------------------------|------------------|----------------------|----------------------|--------------------|
-| **grounding_dino_detect_model**                     | Zero-shot Object Detection (Prompt: 'bark beetle'), Feature extraction | https://huggingface.co/IDEA-Research/grounding-dino-base                          | https://arxiv.org/abs/2303.05499     | (1, 256)                   | 0                          | 0.673            | 0.969                | -0.239               | 0.997              |
-| **yoloworldv2_bb_detect_model**                     | Zero-shot Object Detection (Prompt: 'bark beetle'), Feature extraction | https://github.com/ultralytics/assets/releases/download/v8.3.0/yolov8x-worldv2.pt | https://arxiv.org/pdf/2401.17270v2   | ( , 640)                   | 0                          | 0.004            | 0.007                | 0.686                | 0.577              |
-| **yolov8x_bb_detect_model**                         | Single-class Object Detection, Feature extraction                      | https://huggingface.co/IBBI-bio/ibbi_yolov8_od                                    | https://arxiv.org/abs/2408.15857     | ( , 640)                   | 46 781                     | 0.985            | 0.988                | 0.558                | 0.554              |
-| **yolov9e_bb_detect_model**                         | Single-class Object Detection, Feature extraction                      | https://huggingface.co/IBBI-bio/ibbi_yolov9_od                                    | https://arxiv.org/abs/2402.13616     | ( , 512)                   | 46 781                     | 0.987            | 0.989                | 0.338                | 0.568              |
-| **yolov10x_bb_detect_model**                        | Single-class Object Detection, Feature extraction                      | https://huggingface.co/IBBI-bio/ibbi_yolov10_od                                   | https://arxiv.org/abs/2405.14458     | ( , 640)                   | 46 781                     | 0.990            | 0.987                | 0.428                | 0.559              |
-| **yolov11x_bb_detect_model**                        | Single-class Object Detection, Feature extraction                      | https://huggingface.co/IBBI-bio/ibbi_yolov11_od                                   | https://www.arxiv.org/abs/2410.17725 | ( , 768)                   | 46 781                     | 0.989            | 0.989                | 0.656                | 0.569              |
-| **yolov12x_bb_detect_model**                        | Multi-class Object Detection, Feature extraction                       | https://huggingface.co/IBBI-bio/ibbi_yolov12_oc                                   | https://arxiv.org/pdf/2502.12524     | ( , 768)                   | 46 781                     | 0.913            | 0.755                | TBD                  | TBD                |
-| **rtdetrx_bb_detect_model**                         | Single-class Object Detection, Feature extraction                      | https://huggingface.co/IBBI-bio/ibbi_rtdetr_od                                    | https://arxiv.org/abs/2304.08069     | ( , 384)                   | 46 781                     | 0.885            | 0.890                | 0.386                | 0.534              |
[…]
