### 2025-11-11 6df02de
added evaluation warning and advice
