### 2025-11-11 9feb258
added google colab quickstart notebook
