### 2025-11-11 d9b1758
Feature/v0.2.2 edits (#48)
* fixed mantel's R and removed V-measure

* added genus classification

* attempt 1 for OOD evaluation

* ood evaluation

* added codemeta.json file

* batch processing fix for embeddings eval

* fix all models to run eval

* fix output tables for eval

* fixed yoloworld eval

* fixed groundingdino eval

* added model cards with evaluation summaries

* update model summary table csv

* udpate summary table csv

* embedding adjustment for feature extractors

* Model table update

* docs fix image renbdering

* added custom data evaluation notebook example

* added isn
[…]
