+++ README.md
@@ -3,6 +3,7 @@
 [![PyPI version](https://badge.fury.io/py/ibbi.svg)](https://badge.fury.io/py/ibbi)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 [![Documentation](https://img.shields.io/badge/docs-latest-blue.svg)](https://gcmarais.com/IBBI/)
+[![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/ChristopherMarais/IBBI/blob/main/notebooks/ibbi_quickstart.ipynb)
 
 **IBBI** is a Python package that provides a simple and unified interface for detecting and classifying bark and ambrosia beetles from images using state-of-the-art computer vision models.
 
@@ -61,6 +62,7 @@ The IBBI package provides a powerful, modern solution to overcome these obstacle
   - [Workflow: How the Models Were Built](#workflow-how-the-models-were-built)
   - [Package API and Usage](#package-api-and-usage)
   - [Installation](#installation)
+    - [Hardware Requirements](#hardware-requirements)
   - [Quick Start](#quick-start)
   - [Available Models](#available-models)
   - [Advanced Usage](#advanced-usage)
@@ -74,7 +76,7 @@ The IBBI package provides a powerful, modern solution to overcome these obstacle
 The trained models in `ibbi` are the result of a comprehensive data collection, annotation, and training pipeline by the Forest Entomology Lab at the University of Florida.
 
 <p align="center">
-  <img src="docs/assets/images/data_flow_ibbi.png" alt="IBBI Data Flow" width="800">
+  <img src="https://media.githubusercontent.com/media/ChristopherMarais/IBBI/main/docs/assets/images/data_flow_ibbi.png" alt="IBBI Data Flow" width="800">
 </p>
 
 1.  **Data Collection and Curation:** The process begins with data collection from various sources. A zero-shot detection model performs initial bark beetle localization, followed by human-in-the-loop verification to ensure accurate bounding box annotations. Species classification is performed by expert taxonomists to provide high-quality species labels.
@@ -84,7 +86,7 @@ The trained models in `ibbi` are the result of a comprehensive data collection, 
 
     Note - No additional training is performed for the zero-shot detection and general feature extraction models, as they leverage pre-trained weights from large, diverse datasets.
 
-1.  **Evaluation and Deployment:** A held-out test set is used to evaluate all models. A summary of performance metrics can be viewed with `ibbi.list_models()` or by simply viewing the [`model summary table`](src/ibbi/data/ibbi_model_summary.csv). Alternatively evalutation can be done independently by running `ibbi.Evaluator()`. The trained models are stored under the [`IBBI-bio Hugging Face Hub community`](https://huggingface.co/IBBI-bio) for easy access.
+3.  **Evaluation and Deployment:** A held-out test set is used to evaluate all models. A summary of performance metrics can be viewed with `ibbi.list_models()` or by simply viewing the [`model summary table`](src/ibbi/data/ibbi_model_summary.csv). Alternatively evalutation can be done independently by running `ibbi.Evaluator()`. The trained models are stored under the [`IBBI-bio Hugging Face Hub community`](https://huggingface.co/IBBI-bio) for easy access.
 
 ---
 
@@ -93,7 +95,7 @@ The trained models in `ibbi` are the result of a comprehensive data collection, 
 The `ibbi` package is designed to be simple and intuitive. The following diagram summarizes the main functions, classes and methods including their inputs and outputs.
 
 <p align="center">
[…]
