### 2025-11-11 #8796 @ChristopherMarais
Hello! Thank you so much for this incredibly detailed and constructive review.

Hello! Thank you so much for this incredibly detailed and constructive review. We truly appreciate the time and expertise you dedicated to improving this package. We agree with all of your points and have addressed them in the latest update.

Here is a point-by-point summary of the changes we've made based on your feedback. (The latest version of the package and the paper are published from the latest version of the main branch).

### Typos and Syntactic Corrections (Manuscript)

Regarding the manuscript citations for `Goodwin et al.` and `Sun et al.`, we double-checked the `paper.md` file. The source markdown already uses the correct Pandoc format (`[@goodwin2021mosquito]` and `[@sun2024intelligent]`) to generate parenthetical citations, so no changes were made to the source file.

### Documentation

* **Broken/Huge Images:** This was a great catch. We have fixed all broken image links in the `README.md` and the official documentation by using absolute `media.githubusercontent.com` links. This resolves the rendering issues and ensures the images load correctly on all platforms.

* **Missing Plotting Code:** An example of how to plot and visualize the results are now demonstrated in our notebook `2_inference_examples.ipynb`.

### Content Corrections (User Guide)

* **Custom Dataset Evaluation Workflow:** We have created a new, dedicated tutorial notebook: `5_evaluation_custom_dataset.ipynb`. This notebook:
    1.  Explicitly defines the `list[dict]` structure that the `ibbi.Evaluator` expects.
    2.  Provides a self-contained, runnable example of how to create a mock dataset (using `PIL` to create images and defining annotations) in the required format.
    3.  Runs the `ibbi.Evaluator` on this custom-built dataset, demonstrating the full workflow.

* **"Model evaluation" RAM Crash & `TypeError`:** We have addressed this issue in two ways:
    1.  **Immediate Fix:** The code snippets in the `README.md`, `Usage Guide`, and `3_evaluation_examples.ipynb` now all use `data.select(range(N))` to create a small, safe subset for demonstration.
    2.  **Explicit Warning:** We have added a blockquote warning in the "Model Evaluation" section of the `README.md` and `Usage Guide`. This note explains that the `Evaluator` processes data in memory and recommends using subsets to avoid RAM exhaustion.
    3.  **`TypeError` Fix:** The use of `data.select()` also seems to resolve the `TypeError` you identified when trying to slice the `Dataset` object with `data[:10]`.
    4. A more streamlined automated solution is in the works for a future release.

### Install and Set Up Instructions

* **Hardware Recommendations:** We have added a new "Hardware Requirements" subsection to the `Installation` section of the `README.md` and `Usage Guide`.

* **Pixi Install:** The `pixi` installation command has been corrected.

* **Hugging Face Login Requirement:** This has been fixed. All Hugging Face dataset repositories are now **public**. Users can now access the data without any login required.

* **Colab Notebook:** We have created a new `ibbi_quickstart.ipynb` notebook that combines the key "getting started" steps into a single file. This notebook runs through all major features (model creation, prediction, evaluation on a subset, and LIME explanation). We have added an "Open in Colab" badge to the top of the `README.md` for easy access.

### Other Issues (JOSS Checklist)

* **Seeking Support Guidelines:** We have added a new, explicit **"Seeking Support"** section to the top of our `CONTRIBUTING.md` file. This section directs users to open a GitHub Issue for questions, bugs, or feature requests, as you suggested.

### Further Suggestions

* **Tests & CI:** We absolutely agree. We currently have a CI pipeline for *deployment* to PyPI and enforce code quality and formatting locally with `pre-commit` hooks. We have logged a formal issue to explain our stance on this in more detail so users can see what our roadmap of development looks like.
*
* **Fine-tuning & Serverless:** These are excellent suggestions for the future. We have logged the serverless (e.g., Modal) idea as a feature request. For fine-tuning, we have decided to defer a specific tutorial for now, as our current models are wrappers for libraries which already have excellent fine-tuning guides. We will, however, add a comprehensive `ibbi`-specific fine-tuning tutorial for our v1.0.0 tailored model, which will have a unique architecture.

---

Thank you once again @martibosch for your thorough review and helping us refine the package.
