### 2025-11-13 daf218d
Feature/v0.2.3 edits (#49)
* added fix images docs
