### 2025-11-13 0b2f671
Feature/v0.2.3 edits (#50)
* added fix images docs

* README links update
