### 2025-12-04 #8796 @goldingn
Hi @ChristopherMarais, now that we have two positive reviews we can move onto editorial checks. I'll do mine first then when I can recommend acceptance, the track editors will take over to do the final few things.

Before I can proceed though, I need you to do the list of things at the top here: https://github.com/openjournals/joss-reviews/issues/8796#issuecomment-3614762184
Please note that the author list, order, and affiliations, and the license must match the repo exactly.

And please also address the apparently invalid and missing paper DOIs spotted by editorialbot here: https://github.com/openjournals/joss-reviews/issues/8796#issuecomment-3614763634

Then my text checks etc. should be pretty quick after that
