### 2025-12-11 cef6903
update DOI issues in paper
