+++ paper.md
@@ -37,7 +37,7 @@ The Intelligent Bark Beetle Identifier (IBBI) is an open-source Python package t
 
 # Statement of Need
 
-The accurate and timely identification of bark and ambrosia beetle species is fundamental to forest health monitoring, pest management, and quarantine efforts worldwide [@hulcr2024platform]. Many species are cryptic or morphologically similar (i.e., exhibit high homoplasy), making identification a significant bottleneck that requires highly specialized taxonomic expertise [@kirkendall2015; @hulcr2015morphology]. Traditional morphological identification methods are slow, not scalable for large ecological datasets, and can be prohibitively expensive, hindering rapid responses to invasive species outbreaks and broad-scale biodiversity research.
+The accurate and timely identification of bark and ambrosia beetle species is fundamental to forest health monitoring, pest management, and quarantine efforts worldwide [@hulcr2017ambrosia; @piper2019prospects; @ramsfield2016forest]. Many species are cryptic or morphologically similar (i.e., exhibit high homoplasy), making identification a significant bottleneck that requires highly specialized taxonomic expertise [@kirkendall2015; @hulcr2015morphology]. Traditional morphological identification methods are slow, not scalable for large ecological datasets, and can be prohibitively expensive, hindering rapid responses to invasive species outbreaks and broad-scale biodiversity research.
 
 While computer vision offers a promising solution, a critical gap exists in the available software ecosystem. There is a lack of specialized, accessible, and benchmarked tools for bark beetle identification. `ibbi` fills this gap by providing a robust, easy-to-install Python package that:
 1.  Focuses specifically on the taxonomically challenging and economically critical bark and ambrosia beetle groups.
@@ -50,7 +50,7 @@ By consolidating these functionalities into a single package, `ibbi` provides a 
 
 Automated species identification using computer vision has seen significant progress, with deep learning models achieving expert-level accuracy in various biological domains [@valan2019; @hoye2021]. This progress has been largely driven by two trends: broad-scale generalist models and bespoke academic models.
 
-Large-scale projects, often powered by citizen-science datasets like iNaturalist, have successfully developed generalist models for identifying a wide range of common taxa [@matheson2014inaturalist]. Similarly, comprehensive toolkits like Pytorch-Wildlife offer powerful, pre-trained models, such as MegaDetector, for identifying a broad spectrum of wildlife in camera trap images [@hernandez2024pytorch]. However, these generalist models, while powerful for common categories, often struggle with the fine-grained distinctions required for less common or cryptic species, particularly in specialized groups like bark and ambrosia beetles, where morphological similarities can lead to misidentification [@kirkendall2015].
+Large-scale projects, often powered by citizen-science datasets like iNaturalist, have successfully developed generalist models for identifying a wide range of common taxa [@horn2018]. Similarly, comprehensive toolkits like Pytorch-Wildlife offer powerful, pre-trained models, such as MegaDetector, for identifying a broad spectrum of wildlife in camera trap images [@hernandez2024pytorch]. However, these generalist models, while powerful for common categories, often struggle with the fine-grained distinctions required for less common or cryptic species, particularly in specialized groups like bark and ambrosia beetles, where morphological similarities can lead to misidentification [@kirkendall2015].
 
 This has led to a growing trend of developing specialized tools for specific taxonomic groups of high economic or ecological importance. For example, various research initiatives have focused on creating models for identifying bees to monitor pollinators or for classifying mosquito species to track disease vectors [@buschbacher2020image, @goodwin2021mosquito]. These efforts demonstrate the recognized need for domain-specific tools, but the resulting models are rarely distributed in an accessible, unified software package.
 
