+++ paper.bib
@@ -1,41 +1,51 @@
-@book{kirkendall2015,
-  title     = {Bark Beetles: Biology and Ecology of Native and Invasive Species},
-  author    = {Kirkendall, L. R. and Biedermann, P. H. and Jordal, B. H.},
+@incollection{kirkendall2015,
+  title     = {Evolution and Diversity of Bark and Ambrosia Beetles},
+  author    = {Kirkendall, Lawrence R. and Biedermann, Peter H. W. and Jordal, Bjarte H.},
+  booktitle = {Bark Beetles: Biology and Ecology of Native and Invasive Species},
+  editor    = {Vega, Fernando E. and Hofstetter, Richard W.},
+  pages     = {85--156},
   year      = {2015},
   publisher = {Academic Press},
-  doi       = {10.1016/C2012-0-06455-8}
+  address   = {San Diego},
+  isbn      = {978-0-12-417156-5},
+  doi       = {10.1016/B978-0-12-417156-5.00003-4}
 }
 
 @incollection{hulcr2015morphology,
-  title     = {The morbid anatomy of a pandemic: bark and ambrosia beetle diversity in the age of invasive species},
-  author    = {Hulcr, Jiri and Atkinson, Thomas H and Cognato, Anthony I and Jordal, Bjarte H and McKenna, Duane D},
-  booktitle = {Bark Beetles: Biology, Ecology, and Management of Native and Invasive Species},
+  title     = {Morphology, Taxonomy, and Phylogenetics of Bark Beetles},
+  author    = {Hulcr, Jiri and Atkinson, Thomas H. and Cognato, Anthony I. and Jordal, Bjarte H. and McKenna, Duane D.},
+  booktitle = {Bark Beetles: Biology and Ecology of Native and Invasive Species},
   editor    = {Vega, Fernando E. and Hofstetter, Richard W.},
-  pages     = {249--271},
+  pages     = {41--84},
   year      = {2015},
   publisher = {Academic Press},
-  doi       = {10.1016/B978-0-12-417156-5.00008-9}
+  address   = {San Diego},
+  isbn      = {978-0-12-417156-5},
+  doi       = {10.1016/B978-0-12-417156-5.00002-2}
 }
 
 @article{valan2019,
-  title     = {Automated identification of insect species from images},
-  author    = {Valan, Miroslav and Peri, Eyal and Glinka, Oksana and Zwiebel, Ran and Pupko, Tal and Blank, Lior and Saltz, David},
-  journal   = {bioRxiv},
+  title     = {Automated Taxonomic Identification of Insects with Expert-Level Accuracy Using Effective Feature Transfer from Convolutional Networks},
+  author    = {Valan, Miroslav and Makonyi, Karoly and Maki, Atsuto and Vondráček, Dominik and Ronquist, Fredrik},
+  journal   = {Systematic Biology},
+  volume    = {68},
+  number    = {6},
+  pages     = {876--895},
   year      = {2019},
-  publisher = {Cold Spring Harbor Laboratory},
-  doi       = {10.1101/692336}
+  publisher = {Oxford University Press},
+  doi       = {10.1093/sysbio/syz014}
 }
 
 @article{hoye2021,
   title     = {Deep learning and computer vision will transform entomology},
-  author    = {Hoye, Toke Thomas and Nater, Alexander and Babil, Pavan K and Velli, Eigil de and filmsbo, Rasmus},
+  author    = {Høye, Toke T. and Ärje, Johanna and Bjerge, Kim and Hansen, O. L. P. and Iosifidis, Alexandros and Leitch,  Kirk J. and Mann,  H. M. R. and Meissner,  J. J. and Melvad, Claus and Raitoharju, Jenni},
   journal   = {Proceedings of the National Academy of Sciences},
[…]
