### 2025-12-11 df3d4c2
paper clean up edits
