### 2025-12-11 1d11133
Feature/v0.2.3 edits (#54)
* added fix images docs

* added fix images docs

* README links update

* Stop tracking notebooks with Git LFS

* Convert .ipynb files from LFS to regular files

* update DOI issues in paper

* zenodo release 0.2.3

* paper clean up edits
