+++ paper.bib
@@ -165,7 +165,7 @@
   volume={239},
   pages={122190},
   year={2024},
-  doi={https://doi.org/10.1016/j.eswa.2023.122190}
+  doi={10.1016/j.eswa.2023.122190}
 }
 
 @article{hernandez2024pytorch,
@@ -193,7 +193,7 @@ volume = {11},
 number = {1},
 pages = {13656},
 year = {2021},
-doi = {https://doi.org/10.1038/s41598-021-92891-9},
+doi = {10.1038/s41598-021-92891-9},
 url = {https://www.sciencedirect.com/science/article/pii/S1574954119303280},
 author = {Autumn Goodwin and Sanket Padmanabhan and Sanchit Hira and Margaret Glancey and Monet Slinowsky and Rakhil Immidisetti and Laura Scavo and Jewell Brey and Sai Sudhakar, Bala Murali Manoghar and Ford, Tristan and others},
 }
