### 2025-12-12 3de4e8c
Removed duplicate in .bib
