+++ paper.bib
@@ -187,18 +187,6 @@
 }
 
 @article{goodwin2021mosquito,
-title = {Mosquito species identification using convolutional neural networks with a multitiered ensemble model for novel species detection},
-journal = {Scientific reports},
-volume = {11},
-number = {1},
-pages = {13656},
-year = {2021},
-doi = {10.1038/s41598-021-92891-9},
-url = {https://www.sciencedirect.com/science/article/pii/S1574954119303280},
-author = {Autumn Goodwin and Sanket Padmanabhan and Sanchit Hira and Margaret Glancey and Monet Slinowsky and Rakhil Immidisetti and Laura Scavo and Jewell Brey and Sai Sudhakar, Bala Murali Manoghar and Ford, Tristan and others},
-}
-
-@article{goodwin2021mosquito,
   title={Mosquito species identification using convolutional neural networks with a multitiered ensemble model for novel species detection},
   author={Goodwin, Autumn and Padmanabhan, Sanket and Hira, Sanchit and Glancey, Margaret and Slinowsky, Monet and Immidisetti, Rakhil and Scavo, Laura and Brey, Jewell and Sai Sudhakar, Bala Murali Manoghar and Ford, Tristan and others},
   journal={Scientific reports},
