### 2025-12-15 #8796 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
